Execute a C Function on a New Thread

Creates a new thread on which a loaded shared object symbol (DL FUNC) is executed with 'args' and the results assigned as a special attribute to 'object'. [EXPERIMENTAL - under development - use with exteme caution]

do_threaded(symbol, package, args, object)

Arguments

symbol

a character string giving the symbol name of a loaded function i.e. through a loaded package or by dyn.load.

package

if supplied, the package name as a character value, where native C routines exported via R_RegisterCCallable() will be searched for 'symbol'.

args

an object to be passed to the function.

object

an object (not NULL) on which to attach the return value.

Value

Invisible NULL. The return value of the function will be attached as the attribute "" to 'object'.

Details

Do not pass a function that either calls back into the R interpreter via Rf_eval() or equivalent, or otherwise exhibits non thread-safe behaviour.

The current implementation requires that the function take a single argument 'args'.