Configure Next Mode

Configures 'next' mode by registering 'refhook' functions for serialization and unserialization. This permits sending and receiving reference objects, such as those accessed via an external pointer, between different R sessions.

next_config(inhook, outhook)

Arguments

inhook

a function (for custom serialization). The signature for this function must accept a list and return a raw vector, e.g. safetensors::safe_serialize, or else NULL to reset.

outhook

a function (for custom unserialization). The signature for this function must accept a raw vector and return a list, e.g. safetensors::safe_load_file, or else NULL to reset.

Value

Invisibly, a pairlist comprising the currently-registered 'next' configuration.

Details

Calling this function without any arguments returns (invisibly) the currently-registered 'next' configuration.

Alternatively, calling this function with a configuration pairlist previously returned by this function registers the supplied configuration.

Refhook

The 'refhook' functions are a native feature of R's serialization mechanism and apply to all non-system reference objects (external pointers, weak references, and all environments other than namespace and package environments and the Global Environment).

Examples

cfg <- next_config(inhook = function(x) serialize(x, NULL),
                   outhook = unserialize)
cfg
#> [[1]]
#> function(x) serialize(x, NULL)
#> <environment: 0x55e37e91bb10>
#> 
#> [[2]]
#> function (connection, refhook = NULL) 
#> {
#>     if (typeof(connection) != "raw" && !is.character(connection) && 
#>         !inherits(connection, "connection")) 
#>         stop("'connection' must be a connection")
#>     .Internal(unserialize(connection, refhook))
#> }
#> <bytecode: 0x55e37e966380>
#> <environment: namespace:base>
#> 

nul <- next_config(NULL, NULL)
print(next_config())
#> [[1]]
#> NULL
#> 
#> [[2]]
#> NULL
#> 

print(next_config(cfg))
#> [[1]]
#> function(x) serialize(x, NULL)
#> <environment: 0x55e37e91bb10>
#> 
#> [[2]]
#> function (connection, refhook = NULL) 
#> {
#>     if (typeof(connection) != "raw" && !is.character(connection) && 
#>         !inherits(connection, "connection")) 
#>         stop("'connection' must be a connection")
#>     .Internal(unserialize(connection, refhook))
#> }
#> <bytecode: 0x55e37e966380>
#> <environment: namespace:base>
#> 
print(next_config(nul))
#> [[1]]
#> NULL
#> 
#> [[2]]
#> NULL
#>