Recover an Aio Value

For use on 'ncurlAio' when encountering a character conversion error 'embedded nul in string' (received data is not text data). This function allows recovery by retrieving the value as a raw vector.

recover_aio(aio)

Arguments

aio

an ncurlAio (object of class 'ncurlAio').

Value

The passed object (invisibly).

Details

A raw vector will be available at $data, as if convert = FALSE had been specified.

Note: calling this function on an unresolved Aio is an error and any resulting operation is not guaranteed.

Examples

res <- ncurl_aio("https://shikokuchuo.net/nanonext/reference/figures/logo.png")
tryCatch(call_aio(res), error = identity)
#> <simpleError in (function (aio) .Call(rnng_aio_http, data, response, 3L))(): embedded nul in string: '\x89PNG\r\n\032\n\0\0\0\rIHDR\0\0\002\006\0\0\002X\b\003\0\0\0.e\xde\025\0\0\003\0PLTE\0\0\0;w\xaa=m\xb0?r\xaf?r\xaf?r\xaf?r\xaf?r\xaf?r\xb0?r\xaf=r\xb0?r\xaf?r\xaf>r\xaf?r\xaf>r\xaf>r\xb0?r\xaf>r\xb0?r\xaf>s\xb0?r\xaf?r\xaf?r\xaf?q\xaf?r\xaf>r\xac?r\xaf?r\xb0?r\xaf>r\xaf?r\xaf?r\xaf?r\xaf?r\xaf?s\xaf?r\xaf?r\xaf?r\xaf>r\xaf?r\xaf\027\035A?r\xaf\xff\xff\xff9e\x9d'Ao\xe1\xe1\xe3\027\035B\xef\xef\xf0\xfc\xfc\xfcz{\x84\027\037D>p\xad\xa3\xa3\xa9\xe4\xe4\xe5=n\xaa\031!G<l\xa7\xb6\xb6\xba\xac\xac\xb1\033'N\032$J\xe6\xe6\xe7:i\xa3*Eu\xf7\xf7\xf7\036-V\035*R\xd8\xd8\xda+Iz\xdet\001\xd3\xd3\xd5 0Y\xf2\xb6/\xef\x9b\005\xe3}\001\xe1y\001\xf9\xf9\xf97b\x9b\xf2\xbc<.O\x82\xf2\xba6\x96\x96\x9d0R\x86#6a!3]\xf2\xb3)\xef\x9f\017\xccb\001\xc6^\001\xdd\xdd߽\xbd\xc0\177\177\x886_\x97$9e\xb0S\003\xe7\x80\001\xce\xce\xd13Z\x91&>l\xf0\xa5\022\xb3\xb3\xb7-L~rs}\xf1\xad 1U\x8a\xf0\xa0\a\x91;\003\xec\x93\002\xacN\002\xec\x8f\001\xe9\x85\001\xf2\xf2\xf3\xd9s\003\xb6S\002\xd8m\001\xec\>
recover_aio(res)$data
#>    [1] 89 50 4e 47 0d 0a 1a 0a 00 00 00 0d 49 48 44 52 00 00 02 06 00 00 02 58
#>   [25] 08 03 00 00 00 2e 65 de 15 00 00 03 00 50 4c 54 45 00 00 00 3b 77 aa 3d
#>   [49] 6d b0 3f 72 af 3f 72 af 3f 72 af 3f 72 af 3f 72 af 3f 72 b0 3f 72 af 3d
#>   [73] 72 b0 3f 72 af 3f 72 af 3e 72 af 3f 72 af 3e 72 af 3e 72 b0 3f 72 af 3e
#>   [97] 72 b0 3f 72 af 3e 73 b0 3f 72 af 3f 72 af 3f 72 af 3f 71 af 3f 72 af 3e
#>  [121] 72 ac 3f 72 af 3f 72 b0 3f 72 af 3e 72 af 3f 72 af 3f 72 af 3f 72 af 3f
#>  [145] 72 af 3f 73 af 3f 72 af 3f 72 af 3f 72 af 3e 72 af 3f 72 af 17 1d 41 3f
#>  [169] 72 af ff ff ff 39 65 9d 27 41 6f e1 e1 e3 17 1d 42 ef ef f0 fc fc fc 7a
#>  [193] 7b 84 17 1f 44 3e 70 ad a3 a3 a9 e4 e4 e5 3d 6e aa 19 21 47 3c 6c a7 b6
#>  [217] b6 ba ac ac b1 1b 27 4e 1a 24 4a e6 e6 e7 3a 69 a3 2a 45 75 f7 f7 f7 1e
#>  [241] 2d 56 1d 2a 52 d8 d8 da 2b 49 7a de 74 01 d3 d3 d5 20 30 59 f2 b6 2f ef
#>  [265] 9b 05 e3 7d 01 e1 79 01 f9 f9 f9 37 62 9b f2 bc 3c 2e 4f 82 f2 ba 36 96
#>  [289] 96 9d 30 52 86 23 36 61 21 33 5d f2 b3 29 ef 9f 0f cc 62 01 c6 5e 01 dd
#>  [313] dd df bd bd c0 7f 7f 88 36 5f 97 24 39 65 b0 53 03 e7 80 01 ce ce d1 33
#>  [337] 5a 91 26 3e 6c f0 a5 12 b3 b3 b7 2d 4c 7e 72 73 7d f1 ad 20 31 55 8a f0
#>  [361] a0 07 91 3b 03 ec 93 02 ac 4e 02 ec 8f 01 e9 85 01 f2 f2 f3 d9 73 03 b6
#>  [385] 53 02 d8 6d 01 ec 8c 01 fe fe fe a7 a8 ad 22 27 45 c2 5a 01 c4 c5 c7 35
#>  [409] 5d 94 f1 b0 25 a9 4c 03 bb 56 02 d4 6a 01 32 58 8d 4b 4d 5e ee 96 04 e6
#>  [433] 84 03 eb 88 01 ec ec ed e8 e8 e9 cb cb ce 25 3c 69 98 3f 02 db 70 01 d1
#>  [457] 67 01 a0 a1 a6 99 9a 9f ee c6 6b f1 aa 1c f1 a9 17 c1 5f 02 bd 5a 02 92
#>  [481] 92 99 88 89 91 f3 c7 54 d5 99 51 f3 c5 4a c8 62 03 a4 48 02 39 67 a1 7d
#>  [505] 7e 87 64 65 71 30 33 4c 1d 22 43 a2 44 03 93 42 03 af 4d 02 f0 ca 74 50
#>  [529] 52 61 d4 6d 02 9c 42 02 f4 f4 f4 77 78 81 6b 6c 77 e8 b8 66 c0 93 50 ba
#>  [553] 63 11 e7 89 04 c0 c0 c4 8d 8d 95 8b 35 03 da da dc b0 b0 b5 eb c0 6c 27
#>  [577] 2b 48 de 79 01 c8 c8 cb e5 b2 66 de aa 63 e9 b5 5d 40 42 56 39 3c 52 cc
#>  [601] 68 04 cf 64 00 84 85 8d df a4 58 ed bc 56 35 38 4f ee b9 4c a8 54 07 b9
#>  [625] ba bd 55 57 66 9d 9d a3 59 5a 69 f2 c7 65 ee c0 63 f1 c6 5e eb a8 3d ec
#>  [649] 98 0e 84 30 04 47 49 5b da 96 45 bf 8a 44 cc 86 39 9f 7c 37 d9 8c 35 e4
#>  [673] 94 2e d7 7e 21 ca 71 1c c5 6a 15 df 7f 10 5f 60 6d c9 99 56 f3 c1 42 ec
#>  [697] a8 32 af 5a 0e a7 47 03 5c 5e 6c d6 a2 5e e3 aa 5c f1 be 4b ee b4 43 ac
#>  [721] 85 40 f0 b3 36 e1 8d 25 cb 9f 5e ae 8c 48 e3 9f 43 b1 66 1e e8 ae 55 dc
#>  [745] 85 25 ea 96 1d cf 6f 10 b3 94 52 e4 a4 4c 2b 2e 4a af 7b 35 e9 9f 30 ca
#>  [769] 7c 2c d1 77 1e b7 5b 06 a1 88 46 be 7d 30 bf 71 25 76 27 03 9c 60 1f a1
#>  [793] 4f 07 bc 9e 5d d3 8e 43 a0 6c 25 ca 90 4c 9f 53 10 15 e1 22 eb 00 00 00
#>  [817] 29 74 52 4e 53 00 05 0a e8 f8 ee 68 fb 2e f4 0e 72 dc 5e 35 26 20 82 12
#>  [841] 8b 1b d7 7b 55 3c d0 16 e2 c5 aa be cb 99 92 4e 42 b6 9f 48 b0 a4 aa 9d
#>  [865] bd be 00 00 1d 51 49 44 41 54 78 da ec dd c9 4a 63 41 14 c6 f1 8a f3 10
#>  [889] 35 c6 79 9e 87 2a 0c 77 11 c1 08 09 4a 44 c8 4a 04 cd d2 2c 14 04 5f c1
#>  [913] 8d cb 3c 83 6f db 1d db ee 36 b9 e7 9a e9 96 f7 9c aa ef f7 14 45 fd 39
#>  [937] e7 28 00 50 d9 91 d9 c5 31 05 5e 9b 98 1b 32 bf ad 4f 2b f0 56 6a 64 d0
#>  [961] 7c 3a 9e 57 e0 a7 bd 15 f3 df f8 e2 80 02 ff 64 36 4c b3 b5 25 05 9e 19
#>  [985] 98 1d 32 21 9b 78 2b fa 65 6a d0 90 b6 57 15 f8 62 7a dd 44 19 9e 4c 29
#> [1009] f0 c1 fc 81 f9 ce c2 a1 02 e7 0d 2c 8e 9b 36 f6 67 14 b8 6d 37 6d da 1b
#> [1033] 9d 9b 50 e0 ae b1 4d d3 99 c1 11 3c 11 5c 95 dd 36 84 bb e2 d3 b9 09 5b
#> [1057] d9 53 e0 a0 d4 e4 b0 09 7b 28 6b ad 0b d7 a7 26 6c 23 a3 c0 35 87 0b 26
#> [1081] ec a2 16 e8 0f 57 6f 26 6c 68 16 ff cb 6e 99 39 32 61 a7 d7 79 fd 4f f5
#> [1105] c4 84 ad 4d 29 70 c6 c4 dc a8 09 7b 29 ea af f2 b9 73 d3 0c 09 da 25 8d
#> [1129] 9c 1c 76 53 d1 ad 0a cf d4 13 e1 00 09 da 05 cb 2b 26 ec fc 29 d0 84 ab
#> [1153] 47 f3 09 09 da 29 99 0d ea 51 f0 5c d0 b4 e0 fe dd 84 a5 77 15 08 46 e6
#> [1177] 64 f3 76 a5 a3 e5 2f cf 0c 12 b4 53 96 d6 4c d8 c9 bd fe 5e e9 8e 4c d0
#> [1201] 59 05 12 91 39 f9 3c 17 e8 b6 2a 0f 48 d0 8e 58 25 73 f2 6d 49 77 22 a8
#> [1225] 5d 20 41 3b 20 35 49 e5 e4 d7 b2 ee 54 9e fc 5f 3e 42 82 96 84 cc c9 ef
#> [1249] b5 40 77 a1 f8 82 04 2d 1a 99 93 cf 2e f3 ba 4b 95 1b 24 68 b1 b2 3b a3
#> [1273] 64 4e d6 dd 0b e8 04 bd ac 80 b9 d4 08 99 93 2b ba 37 85 67 24 68 81 f6
#> [1297] c8 9c 5c 0f 74 0f 90 a0 85 8a c8 c9 05 dd 97 fb 8f 04 8d 11 27 21 1a d3
#> [1321] c9 64 4e ee 57 80 04 2d c7 14 99 93 ab 3a 0e a5 5b 43 38 c0 88 13 37 cb
#> [1345] 5b 11 39 39 26 e5 57 2a 41 e3 7f 99 95 f9 e3 88 9c 1c 9f a0 86 04 cd 1b
#> [1369] 3d 88 f4 78 a5 e3 45 27 e8 7d 24 68 1e a2 72 72 fc 8a 54 82 1e dd 41 82
#> [1393] 4e de 18 95 93 cf 72 81 b6 82 4e d0 f8 5f 4e d8 ea 76 44 4e b6 a6 4e 26
#> [1417] 68 8c 38 25 80 18 44 22 72 b2 25 05 3a 41 e3 7f 39 29 bb e9 88 41 24 bb
#> [1441] e8 04 3d 84 04 9d 88 99 fd a8 9c 6c 5f 95 4c d0 18 71 ea 14 c7 9c ac 63
#> [1465] 4b d0 5b 48 d0 ed f1 cd c9 fd 27 68 6c d9 4c c0 de 0a 99 93 b5 3d 18 71
#> [1489] 62 27 b3 11 5f 4e 46 82 16 2a 62 10 a9 a8 93 10 e4 ce a8 04 8d ff 65 9a
#> [1513] 8c 9c dc 5b 82 c6 96 cd 1f 37 4d e6 e4 5c 5e 27 a8 fc 8a 11 27 fb da e7
#> [1537] e4 92 4e 16 3d e2 94 c6 88 d3 17 e2 72 72 8f 09 1a 5b 36 7f c6 52 9a cc
#> [1561] c9 81 66 01 09 fa 3b ec 06 91 5a 21 41 0b 12 99 93 59 a9 5f 60 cb 66 88
#> [1585] f0 9c dc 35 6c d9 24 d8 df 6b c9 e4 51 d0 ac 88 11 a7 66 2e e4 64 24 68
#> [1609] 26 22 f7 5a f2 45 6f d9 dc c2 88 53 fc 7b 2d 79 2b 21 41 37 58 de 6b a9
#> [1633] f9 43 82 76 30 27 f7 20 20 13 74 1a 09 3a 9e 9c cc e1 e7 f8 0b 6c d9 b4
#> [1657] 6b 8a 1c 44 aa 6a 49 90 a0 fb de 6b c9 2f 27 23 41 ff 2c fa 4c e2 29 b7
#> [1681] 9f 63 6c d9 24 58 cf c9 fc 7e 8e 91 a0 29 56 f7 5a 72 c9 c9 7f 61 cb 26
#> [1705] c1 a7 9c dc 80 2d 9b 56 65 77 0c e1 4e e2 a3 a0 05 b6 6c 4a 1a 44 6a 86
#> [1729] 04 9d 00 3a 27 d7 05 3f 0a b0 65 93 c9 5e 4b 5e 70 e8 d1 bd 9c dc 21 1c
#> [1753] 7a 74 3d 27 f7 f6 bf 8c 43 8f dd ed b5 74 e6 51 d0 ac 8c 04 dd c5 99 44
#> [1777] a7 1e 05 38 f4 e8 7a 4e 46 82 16 b5 d7 92 17 fa d0 a3 a7 23 4e 11 39 d9
#> [1801] d1 47 01 46 9c 48 ab 07 12 06 91 ac 09 b0 65 f3 cf 99 44 19 83 48 f6 14
#> [1825] 70 e8 91 ce c9 3c 07 91 5a 20 41 c7 65 4c d8 20 92 3d 15 7f 47 9c 22 f6
#> [1849] 5a ba f6 73 8c 2d 9b de e6 64 1c 7a 14 b5 d7 92 17 ef 12 74 c6 83 9c 8c
#> [1873] 2d 9b ed cf 24 fa 90 93 71 e8 51 d4 5e 4b 5e 3c 19 71 f2 2b 27 e3 d0 23
#> [1897] 69 fe d8 b3 9c 8c 43 8f a2 f6 5a f2 e2 f4 a1 47 5f 73 32 b6 6c 3a 3f 88
#> [1921] 14 82 04 2d 7f af 25 2f ce 6d d9 14 b3 d7 92 17 c7 0e 3d 4a da 6b c9 8b
#> [1945] 43 87 1e a5 ed b5 e4 c5 91 2d 9b c8 c9 48 d0 8d 41 24 e4 64 ef 0f 3d 22
#> [1969] 27 63 cb a6 ec bd 96 cc 88 4d d0 ac ce 24 ca 27 74 c4 09 39 f9 83 df 09
#> [1993] 9a e3 99 44 f9 84 6d d9 8c d8 6b 89 47 81 4f 5b 36 91 93 09 de 6d d9 e4
#> [2017] 7d 26 51 be 88 04 cd eb 7f 19 39 d9 be ca 03 f3 2d 9b d9 6d 57 f7 5a f2
#> [2041] c2 3a 41 23 27 b7 e3 c3 96 4d e4 e4 0e b8 7e e8 91 ce c9 d7 78 14 34 71
#> [2065] 3c 41 7b b2 d7 92 17 6e 87 1e 3d da 6b c9 4b 81 d3 96 4d b9 67 12 e5 63
#> [2089] 93 a0 bd db 6b f9 8b bd fb 67 69 2b 0a e3 38 7e 5b 9c 3a 14 5a 4a ff 2d
#> [2113] 16 3a 3d 9c 56 4c 8b 4a 0d 0a 8a 25 53 96 8e b9 5d 2c 88 6b 16 95 4c 75
#> [2137] 88 2e 3a e9 66 e2 22 a8 8b 52 11 07 05 41 f0 cf d6 bd 93 af a1 6f a1 62
#> [2161] ff 78 4c ee 49 6e da a2 c9 f3 7c 3f af e1 07 39 b9 cf 39 bf a7 bd b4 c7
#> [2185] a2 47 93 bd 96 ed a5 0d 46 d0 1a d6 24 76 be e4 96 cd 17 0f a3 a6 58 93
#> [2209] a8 ca df 8c a0 e9 b5 d4 e7 b6 46 d0 77 74 ad 49 ec 7c 6f df 0f b4 d0 b2
#> [2233] c9 9a 44 b5 d2 b6 6c 32 4e 56 2e dd a2 47 d6 24 aa 77 33 8b 1e c3 bd 96
#> [2257] 1c 0a da c3 0d b5 6c be 54 be 26 b1 f3 35 6c d9 64 4d a2 1d 81 45 8f 8c
#> [2281] 93 8d 09 2d 7a e4 21 92 31 81 45 8f cf e9 b5 34 66 f4 53 72 cb 26 bd 96
#> [2305] b6 84 16 3d fe d3 38 99 87 48 9d 27 34 82 a6 d7 d2 98 a1 40 cb e6 ff 1a
#> [2329] 27 0f d8 58 93 d8 f9 46 03 2d 9b f4 5a da 12 5a f4 c8 9a 44 63 42 8b 1e
#> [2353] 59 93 68 4c 68 04 cd 9a 44 63 02 4f 9c e8 b5 34 26 d4 b2 c9 9a 44 63 42
#> [2377] 2d 9b f4 5a 1a 93 f8 c4 a9 eb f1 5d d6 24 1a d3 42 cb e6 bd 07 ac 49 54
#> [2401] 2b b4 e8 91 71 b2 31 a9 5a 36 ef 27 8f 93 f9 93 a8 48 e8 89 53 b3 71 32
#> [2425] 5f 8e 75 09 2d 7a a4 d7 d2 98 06 2d 9b 8f ba e9 b5 b4 23 b1 65 b3 eb 49
#> [2449] 14 3d e3 cb b1 29 89 2d 9b 4f 23 c6 c9 d6 24 8c a0 bb 23 c6 c9 e6 24 8c
#> [2473] a0 23 7a 2d 0d aa 1b 41 47 1c 0a 4c aa 19 41 47 8c 93 6d f2 bf 2f fb 31
#> [2497] 78 c7 38 d9 96 a1 cf 49 31 78 25 b0 e5 35 31 00 31 00 31 00 31 00 31 00
#> [2521] 31 00 31 00 31 00 31 00 31 00 31 68 77 67 ab d2 2a 62 a0 4f 5c 9e ae 4a
#> [2545] 6b 88 81 3e b1 73 eb 83 b3 d2 0a 62 a0 4f ec 2e e4 4b df 24 3d 62 a0 4f
#> [2569] ec 2e 65 72 1b 92 16 31 d0 27 76 bf 15 b6 25 1d 62 a0 4f ec ae cc 2d ec
#> [2593] 4a 0a c4 40 9f ac f3 15 77 f6 a5 29 62 a0 d0 e4 7c c6 79 46 e2 13 69 82
#> [2617] 18 a8 b4 76 5c 74 9e cc fc a6 34 44 0c 94 aa f6 0d 3b cf d4 d8 a1 34 40
#> [2641] 0c f4 5a 1e 9b 72 9e fe c5 55 09 21 06 9a 6d ee e5 9d a7 9c 3d 92 64 c4
#> [2665] 40 b7 d9 89 a2 f3 ac 1f af 49 12 62 a0 dd ee 69 bf f3 e4 f7 2a 52 8f 18
#> [2689] 18 70 98 f3 0f 09 99 dc 1b a9 45 0c 4c a8 94 f2 ce 53 58 92 eb 88 81 11
#> [2713] fb 3b bd ce 33 d7 f7 5d 3c c4 c0 8c 99 c5 1e e7 29 4e 7c 95 3f 88 81 25
#> [2737] 2b b9 8c 7f 5a dc 3a 97 5f 88 81 2d e7 5b 23 fe 69 71 7c 52 2e 11 03 6b
#> [2761] f6 b3 65 e7 f9 b2 2c 17 88 81 3d ab 1f 0a ce 33 7c 3a 43 0c 6c da 18 f7
#> [2785] 0f 09 bd d9 23 62 60 d3 49 bc ee ae 7c 24 06 56 1d 4c 97 89 01 2e 2c 15
#> [2809] 88 01 7e de 57 23 06 20 06 90 6d 7e 14 cc 3b e2 88 08 fe 30 82 cf 47 a8
#> [2833] ff 98 bc 2b 42 0c 8c 39 60 b4 04 06 cd a8 bd 76 52 e2 da 89 3d 5c 42 43
#> [2857] fd 95 d4 aa 78 88 81 09 95 d2 88 f3 f4 70 41 dd a0 c3 6b 47 82 a9 dc 8a
#> [2881] d4 22 06 da f1 78 0d 3c 65 45 dd c3 f6 de ec 81 24 23 06 7a d5 d5 5c cc
#> [2905] 48 08 31 50 aa ba 30 ec 7c 94 de 18 b4 36 58 74 ad 7c 34 26 06 0a 6d d6
#> [2929] 14 e2 6d 9d 49 13 c4 40 1f ea 31 41 59 2e a8 ce 06 45 fa 48 58 ab 51 91
#> [2953] f4 88 81 3e 2c d9 01 2b b7 c0 02 3e b4 11 62 00 62 00 62 00 62 00 62 00
#> [2977] 62 00 62 00 62 00 62 00 62 00 62 f0 83 5d 3a 26 00 18 04 00 18 26 68 0e
#> [3001] 76 20 02 6b 88 c0 26 b5 c0 4d a2 21 68 80 06 68 80 06 68 80 06 68 80 06
#> [3025] 68 80 06 68 80 06 68 80 06 68 80 06 68 40 34 20 1a 10 0d 88 06 44 03 a2
#> [3049] 01 d1 80 68 40 34 20 1a 10 0d 88 06 44 03 a2 01 d1 80 68 40 34 20 1a 10
#> [3073] 0d 88 06 44 03 a2 01 d1 80 68 40 34 20 1a 10 0d 88 06 44 03 a2 01 d1 80
#> [3097] 68 40 34 20 1a 10 0d b8 b5 67 96 06 8f fb bf 0c 0d 1e a7 01 1a a0 01 1a
#> [3121] a0 c1 61 e7 fe 5d 13 87 c3 38 8e ff 3d 0f 1f 10 bf 05 1d 14 84 40 20 53
#> [3145] 96 8e da c5 41 5c 5d 54 dc 1c d4 45 27 dd fc b1 08 d6 45 51 a4 83 42 41
#> [3169] a8 ba b9 77 ba bf e1 fe 85 7b a2 de 59 35 cf 25 ed 25 e6 0a 79 4d 1d ce
#> [3193] fb 3e fd fa 6e 4c 48 da 50 98 41 28 cc 20 14 66 10 0a 24 83 c5 70 4d 5f
#> [3217] b1 de 54 a5 17 fa b8 2a ad 87 d5 3e 09 ee 3c cb 60 b9 d9 97 49 f6 5f 67
#> [3241] d0 4a b2 2a 1d 0c 93 b3 02 00 33 9a 6d ad c8 ad c1 d3 3c 9b 1a 29 30 b3
#> [3265] a6 e7 f2 24 f0 76 55 b6 6c 36 ea 26 98 96 7a 69 ad 49 70 9f 59 7a 39 3d
#> [3289] aa 00 14 0b e9 37 57 3b f0 33 69 29 2d e8 da 2a 69 79 bc 77 06 11 b0 3c
#> [3313] b1 dd 8b c2 1f 85 26 b9 d2 79 29 e0 52 6a 4b 02 ef 56 65 e3 86 c2 07 66
#> [3337] e5 07 09 fc 9f a5 1d b9 dc 81 16 39 9b c0 92 a0 6b 15 b0 62 27 90 0c 9e
#> [3361] f9 8b 98 86 0b 46 99 9c 94 b7 29 d8 98 55 49 e0 c9 aa 96 7e 46 e1 8a 99
#> [3385] 13 5e ea f7 2c bd 34 ae a5 9d 9b ec d7 c1 54 9e 2e bd 16 c1 2a 44 b2 0c
#> [3409] 44 8d 7f cd 60 50 c1 35 83 04 8e 03 d5 f7 24 f0 62 55 f6 23 0e 1b 69 21
#> [3433] 3f 7f 67 e9 68 b8 35 ea 92 93 bc 02 8b 97 2f e3 a8 81 d5 d7 c1 64 f0 40
#> [3457] 59 30 a5 97 a6 dd 76 c9 d0 70 90 24 07 5b 1c d4 8c c4 e3 6b fe 7d fc d4
#> [3481] ac 44 85 fd f3 70 55 d6 3b 2d 33 ca 6e 5f c7 e3 4e f3 f7 4b eb 43 12 f8
#> [3505] 37 cb 54 c1 a2 66 c9 a7 de be d7 2d e9 0a 16 ed 99 9c e4 60 c9 dd 7e 54
#> [3529] a8 07 0a 26 83 ee 1b d8 ec 9d 8e aa 15 58 b4 a5 d3 0f 25 80 7a 62 4c 67
#> [3553] 83 ad 06 cb 2b 09 3c 58 95 aa d1 63 04 a5 01 9d fc 9c 9b b0 c4 d7 24 f0
#> [3577] 6b 96 67 13 ac 98 7d 3f 37 6a c0 12 ad 92 83 41 04 4c 8d e9 ec 41 81 25
#> [3601] 28 a0 0c 72 45 00 f3 9b 52 93 e4 a0 90 ee d8 1f eb 74 12 78 b1 ea cc e6
#> [3625] 23 60 97 3a ed 83 c0 a7 59 56 51 30 d5 a2 8f 92 45 b0 2c 39 e9 69 60 91
#> [3649] c1 f9 f2 a1 7e fe 9c 90 fd 78 3a 7a 01 ab 3d 7d d4 fb b7 0c d4 cd 77 9c
#> [3673] 06 8b 93 83 77 e9 04 56 2d e9 cc e3 55 63 b0 cc d6 74 a1 1a 87 65 4a 02
#> [3697] 7f 66 99 80 a9 0e 5d 4a c2 d2 25 27 cd ab d4 de c0 cc 31 b1 40 ae 14 58
#> [3721] 86 2e 4c c1 8a 43 fa bc 8d 02 6b d3 99 b7 ab ae 46 60 d1 d5 cd ba 1a 58
#> [3745] 7d 40 02 3f 66 d9 09 07 71 1d 4c 27 47 06 98 d9 bb 38 90 96 28 c0 0c 52
#> [3769] 3f e9 42 59 13 de 4c 67 f1 eb ad f1 76 d5 39 2c 53 e9 28 d1 22 81 1f b3
#> [3793] 4c 84 f2 de 15 00 b5 21 27 d5 02 58 9a 0e 16 29 b0 19 05 99 41 de 36 e8
#> [3817] 04 7d c1 8b f0 21 ed d1 aa 35 f1 27 2d 75 dc 53 81 0f b3 2c 46 60 31 ba
#> [3841] a5 83 cd c9 51 1b 96 d2 f9 6d d5 76 41 66 a0 db 9f 8f 66 e9 0b 72 c2 fb
#> [3865] e4 cd aa 79 58 1e c9 46 52 3a 8c fb 35 cb 2b 98 f6 93 6e 3d 82 e9 6e 7f
#> [3889] 66 b4 1f fc d5 58 81 b5 28 c8 0c 3a 74 ad 04 96 a6 2f 48 0a af f4 66 d5
#> [3913] 04 d8 a8 4c 36 f6 0a ac 49 02 ef 67 c9 88 6f 76 15 4c 23 67 ab df 07 b7
#> [3937] 41 e4 58 5d 90 19 68 0b ba d6 16 17 91 f5 f7 9b dd 6e 97 14 4e b1 3d 58
#> [3961] d5 e1 ec 2b 0e f6 42 02 ef 67 49 cb 87 fe 1a d8 8e 9c 75 4f ed ce c1 0a
#> [3985] fb 40 33 d0 85 03 5e 8d dc e9 77 72 d9 48 c1 c4 91 90 81 37 ab 46 c1 12
#> [4009] f2 11 16 11 12 78 3f 4b 01 6c 2a 17 d2 21 17 de c0 b4 27 13 6c 4a 81 66
#> [4033] 30 b1 a9 14 2c 4a 2e ac b7 b3 53 00 42 06 5e ae ba 28 42 de ae 12 58 81
#> [4057] 04 9e cf d2 87 25 2f 5f 0a 36 c9 85 45 1c 4c 81 55 28 d8 0c 72 74 e3 c1
#> [4081] 65 06 ab c4 08 cc 21 03 cf 56 7d 87 e5 81 6c 3d 82 15 cb 24 f0 67 96 e6
#> [4105] a3 9d 19 d8 9c dc 18 9b 38 aa ad 02 ce a0 f4 e5 0c a6 51 fc a1 b4 42 ad
#> [4129] 5e af 17 84 0c 3c 59 35 0f 4b 8f 6c 75 60 59 92 c0 fb 59 1c 24 dc 9f 53
#> [4153] 33 f5 40 01 67 10 fb 6a 06 89 22 0e 54 3a d7 ee 0d e8 a0 29 64 e0 c9 aa
#> [4177] 0f b0 54 c9 d6 33 2c 1b 12 78 3f 8b 83 37 72 27 0a 8b 41 df 35 83 0c 0e
#> [4201] 6a b1 25 59 fc cd c0 f9 9d ee 0a 8d f8 36 8b 83 1c b9 52 c2 81 7a fe a6
#> [4225] 19 c4 60 51 a5 32 d1 bd 32 18 c3 32 26 5b 53 58 56 24 f0 67 96 b8 ac 4d
#> [4249] 6e f4 4c 1c d5 fb df 32 83 8d 06 56 e0 88 ef 97 c1 10 96 27 b2 d5 02 33
#> [4273] e9 5e 19 54 a5 f3 14 99 fc d8 01 d8 e4 5b 66 30 01 33 9f e9 4a d2 cf 0c
#> [4297] c8 04 db 92 ad 9c 70 a5 ef eb 2c af e4 9a 3c 74 fc 78 91 f3 fa 0d 33 28
#> [4321] 8f ec 3f ff 32 be 66 50 87 7c 81 6d 80 e9 f7 cb 20 25 26 e9 de b3 3a 3e
#> [4345] 99 9a 05 8b 2e bf 5f 06 5d 30 73 4f d7 74 5f 33 c8 82 45 e4 33 6e 64 ee
#> [4369] 97 41 c3 f1 5e aa cb 07 94 df 88 aa 23 b0 c6 f7 cb 40 b8 f5 d2 37 7d cd
#> [4393] a0 09 a6 86 f2 75 7c e7 7e 19 6c a5 7f e1 de e4 cf b9 61 0b 96 f6 27 33
#> [4417] a8 07 9e 41 c2 fe 96 58 0b be 66 30 54 e2 a5 58 05 4c 5b dc 2f 83 bd fa
#> [4441] 74 77 76 bf 96 50 ec 9e 8f a2 a3 21 09 84 33 a1 c0 33 78 b3 3d 8a 95 6b
#> [4465] fe 66 40 3a 58 b4 4f 37 aa 9a 74 da e0 db 2c 06 58 84 be 6c 19 fd 30 f2
#> [4489] 4e 03 33 48 20 dc 41 09 3c 83 a4 ed 33 53 73 1c a4 7c db fa 0e 2c 19 e1
#> [4513] ac a1 38 be 67 06 0f b0 34 e9 ab b2 17 8f 55 c6 3e f7 bf 35 c1 8a fd a0
#> [4537] 33 68 83 69 57 63 b4 15 0e 6a be 6d 3d a5 c1 54 d7 f6 be 12 b2 74 cf 0c
#> [4561] c8 b0 99 c5 bd c7 ab 93 99 f4 e9 51 24 77 3a b0 e4 83 ce 60 af 6e 9f e4
#> [4585] 9f 9a 80 1a 01 28 f8 b7 f5 79 05 a6 75 ed fa d3 36 f7 cd e0 87 06 36 7a
#> [4609] a5 1b 55 72 b4 b9 be 38 78 37 3f f3 c8 57 15 96 5c 90 19 9c e3 35 1f e8
#> [4633] 8f 72 ae 68 0d 96 06 a0 fc db 7a 4a c2 62 c6 e8 6c 31 57 0e cf 25 fb 34
#> [4657] 4b ab 08 a6 72 7d fa 68 d0 8a 47 c8 91 7e f3 c0 51 12 96 e4 67 ee 48 15
#> [4681] 96 41 67 d0 39 bd 1d e5 53 04 8f 75 30 bd 6c 80 0d fd db 7a 7a c1 41 bc
#> [4705] 75 da 82 7d b3 ee 70 67 d7 bf 59 72 38 28 24 37 74 d2 3f fe b2 7f cf dd
#> [4729] 1d a5 47 e9 f7 d9 9c 4d 60 89 e7 e9 e8 7d fe 1a 48 06 64 e0 20 fa 52 6a
#> [4753] c5 12 86 06 4b 64 45 0d b0 27 1f b7 7e d0 c0 91 4a 1b 95 8a 11 51 e2 8d
#> [4777] 5d ff 33 a0 79 11 47 f5 6c 25 f1 56 c9 d6 95 bb 24 7b a6 cd 85 c1 58 81
#> [4801] c5 17 e4 46 4f fd 5e 78 32 79 d1 a3 40 26 98 0c 56 71 5c 6b f4 4f 57 92
#> [4825] 4d 3f b7 9e e6 0a 37 cc 2d 05 91 01 b5 35 d8 89 d0 5f 95 23 b6 e7 32 b9
#> [4849] cf f4 9c c0 a5 59 30 19 d0 d2 c0 85 da 94 d8 16 6c e2 eb d6 53 3e 8d 2b
#> [4873] fa 3b 05 93 01 0d 1b 0a d7 52 a5 b5 9b b7 b0 79 53 47 ea 13 0f 22 95 2b
#> [4897] b8 30 ba 5f 06 f2 df 7b 51 e9 56 f9 fc 38 46 da d7 ad 67 4f 59 13 7f 68
#> [4921] 8d 3c 39 f0 73 96 de 24 8a b3 62 3c 91 77 75 47 09 33 e9 3a a8 b6 26 57
#> [4945] a6 69 9c 69 d9 be cb 0c 7c d0 8b 55 b2 ba d1 48 b4 f7 74 67 3f 9f 92 15
#> [4969] 43 d7 8d 4a f2 a1 4c 41 1b c7 32 59 9e a5 31 89 75 57 74 47 bb 56 a6 61
#> [4993] cc f4 6c 26 96 1f 84 7f 17 31 14 fe 79 cc 50 98 41 28 cc 20 14 66 10 0a
#> [5017] 33 08 85 19 84 c2 0c 42 61 06 a1 30 83 50 98 41 28 cc 20 14 66 10 0a 33
#> [5041] 08 85 19 84 c2 0c 42 bf d8 a5 03 01 00 00 00 00 41 fe d6 0f 52 0c 69 80
#> [5065] 06 68 80 06 68 80 06 4c 03 a6 01 d3 80 69 c0 34 60 1a 30 0d 98 06 4c 03
#> [5089] a6 01 d3 80 69 c0 34 60 1a 30 0d 98 06 4c 03 a6 01 d3 80 69 c0 34 60 1a
#> [5113] 30 0d 98 06 4c 03 a6 01 d3 80 69 c0 34 60 1a 30 0d 98 06 4c 03 a6 01 d3
#> [5137] 80 69 c0 34 60 1a 30 0d 98 06 4c 03 a6 01 d3 80 69 c0 34 60 1a 30 0d 98
#> [5161] 06 4c 03 a6 01 d3 80 69 c0 34 20 76 c9 58 57 41 18 0a c3 cf d7 2e 84 a5
#> [5185] 5d 1d 3a 34 c4 e8 e2 d6 89 eb 6c bb 18 f7 bb b1 32 28 93 86 09 d4 37 c0
#> [5209] 45 88 93 a3 6f 70 3f d0 9b f0 0a 26 7c f4 1c ce 39 fd a1 90 73 60 1a 83
#> [5233] 09 98 c6 60 02 be 6e 0c 7e 47 6c 20 cf 37 23 f2 62 0f 45 96 d5 40 fe d1
#> [5257] 0d ae aa 2a 2a 55 5d 57 f5 99 18 ce 00 a3 70 0b 6d c9 d5 b5 6d 87 f5 44
#> [5281] 51 1c 45 8f 9e 68 44 4c 11 1a 82 1e a1 44 2c 44 ac ac b7 d6 04 1f 9c 72
#> [5305] 4a 39 21 14 cb 1a a9 bd 72 4e a8 41 e9 90 a2 f7 c6 68 6d 12 6d ac 12 2c
#> [5329] 61 83 55 36 c8 e0 ad 73 8a 5c 20 e3 7d 4a e1 93 10 50 07 6f 80 0a 1b 1c
#> [5353] 86 71 48 d4 ee 60 8b 35 71 b3 8b 15 5f d4 3f 1b 8c b7 42 38 64 0d 5b fc
#> [5377] 15 ae f7 71 84 60 08 df 0e 08 9a a8 41 46 fa 35 63 90 e5 03 ff b7 6c bf
#> [5401] cf b2 02 32 62 b2 fb e1 70 3c 1e ef f7 e7 b3 78 16 45 9d e7 75 be a9 07
#> [5425] 4e e5 e9 54 14 18 6e a0 2c 59 f0 49 ca ae c3 e0 72 b9 dc 5e 37 78 3d 06
#> [5449] 94 b3 46 5f af 5a 6b eb ac f5 be 37 df 77 6b 80 52 a0 ed 21 31 81 b6 ca
#> [5473] f9 3c 49 16 8b 54 a6 52 26 f0 76 8b f5 72 3d 9f cb 84 3e 7a af d9 31 08
#> [5497] 29 ce 66 3f cb d5 22 d5 89 94 5a a7 04 72 bd 5c ad fe a8 29 9f d7 a6 c1
#> [5521] 38 0e ff 77 01 f3 06 c4 8b 49 4e cb c1 83 48 7c 43 4c 0e 05 6d 20 c9 29
#> [5545] 39 74 f6 50 7a 13 21 17 67 8e 65 1e 5a 7a f0 b8 c2 2a 6e 55 86 93 0a da
#> [5569] 6e 6e d6 62 ad b8 c3 64 f8 7c 2b 8a ff c2 9e e6 c7 fb be f9 e6 e5 4d 3f
#> [5593] 4f 62 9a be 49 0d 55 84 6f ca c9 4e fd 34 d5 b1 f6 b5 ef fb a9 89 3a ad
#> [5617] 87 77 4d 24 32 d3 14 2d 10 e6 26 4a b0 b0 87 2d bb c5 4e b9 4e ef c9 9c
#> [5641] b6 bd 8d 7e b7 88 5e f2 bf 75 fb c6 1d 9b 32 ea f1 43 5c c6 13 a4 fb c4
#> [5665] 43 e0 28 2e 5c 1b 0d 76 e1 99 ec c2 2e b1 93 f9 10 3e ec 7e a0 f1 62 3a
#> [5689] 9d 76 60 38 5c 2e 97 67 67 67 e8 b1 8b 25 67 b0 3c dd 3b dd b0 5c 9e 2e
#> [5713] 89 1a 48 7a b5 da db db e4 ff 0e 48 7f 75 20 cc 66 5f a0 66 87 91 6d ea
#> [5737] 3a 0f 82 a0 d6 23 d3 4c d9 7c 5f d7 a4 fa 17 c2 d0 3a 8e 75 9c 29 cf 0b
#> [5761] 6a 65 19 9c 69 e5 79 4c fe b1 ce 95 65 29 45 5f 52 d4 5c 88 e3 80 a2 28
#> [5785] 29 cb a8 b0 2c 8f 6e 80 13 86 0a 32 c3 e0 a8 2c 65 79 59 96 cb 04 7e ae
#> [5809] f9 79 96 e7 29 8b 52 60 34 cb 75 ea b3 20 4f 29 15 c4 e2 17 9f 20 3f f7
#> [5833] 91 30 8e fd 9a 5a 04 d3 7e 5e d7 7e 9c 8a 9f 77 b6 f9 86 f0 25 49 47 3a
#> [5857] c0 23 04 19 a5 f6 b6 78 64 8f 7c 0c 4f d3 11 6c b7 5a d7 46 83 17 ff 33
#> [5881] 1c 92 fa b4 73 d4 f9 cb d1 9c f6 55 a7 b3 2f 48 e2 7f 24 39 95 ee eb 7f
#> [5905] ec bf 3e 00 98 9d cf 66 b3 83 83 15 4a ac 66 70 7e 7e 7e 08 eb c3 c3 b7
#> [5929] 82 da 1c b3 ba 0e 94 84 d0 b3 48 a7 8e e3 dc 0b 24 69 72 a0 9d b3 07 ca
#> [5953] 10 0a 88 0a 23 2a 07 65 92 44 45 61 18 16 9b a5 8c 2a 4a a2 04 0a 6a 18
#> [5977] b6 94 a7 8a c4 69 b7 9b cd 86 eb 94 49 64 84 86 b2 8a a8 4a 1c 07 33 a2
#> [6001] b2 ac aa 82 4a a6 ec 29 d5 b3 8c 30 8c b8 c8 9c 91 cc 4a 37 34 30 a5 57
#> [6025] 94 65 49 a1 57 e3 60 9c 65 46 8f 25 86 21 5b 15 46 52 81 7b ac 5d 3e 22
#> [6049] 5a 9b 77 6c dc cd 3c d5 53 5e 86 63 5c 4a e9 d7 99 a0 7a 34 28 1d 8d ae
#> [6073] 8d 06 73 98 02 ef fd d1 7c 7e 24 5c ed ec 8c 77 d8 c7 72 14 ae ae 7e 81
#> [6097] e4 bd df 01 a4 f8 f5 0a be 5f 7c 07 20 6c 80 f5 64 b2 5e 1f 12 3e 06 ac
#> [6121] d7 f4 de 08 cf 9f bf 59 4c 16 8b 45 1f ba fd 2a ec c9 3f 5b 12 97 84 4b
#> [6145] ae 05 ef 31 2d 42 21 65 fa 06 2f f6 c0 11 5c b7 e1 36 9a 0f 1e 34 1b 0d
#> [6169] 77 00 32 3a 60 44 46 9b 0f 9a 6e 97 0a d7 49 08 d5 69 3e 7d f2 e8 c9 93
#> [6193] f7 4f a9 75 06 72 63 a3 d1 e6 ce 46 93 b3 db e5 87 21 dc 5c 25 55 e9 70
#> [6217] 91 51 06 06 83 7e 7f e0 74 dd 7e d7 e9 76 9d c7 6d 26 a6 a4 10 31 92 aa
#> [6241] 2a 93 92 5a a1 bb 81 45 86 9e f8 4b cc a2 4a c0 83 44 49 15 a1 49 c5 63
#> [6265] 61 18 03 e2 4d b5 60 28 7c 0b d7 46 83 cb cb cb 39 f1 8b 00 3b f0 55 38
#> [6289] 39 f9 08 c2 09 8c 3f 8e c7 af fe 31 06 f8 f6 ed e7 cf 8b cf 70 01 13 c9
#> [6313] 7b 32 59 90 38 d0 43 86 cd c5 e3 97 2f b7 5e 0a c7 c7 3f da ed ad fb 5b
#> [6337] 8f 1e 6d b5 7f b3 53 c6 ba 6d c3 50 14 fd 4d 6b e2 07 10 1c ca 49 03 dd
#> [6361] 99 c8 44 4e e2 c4 12 28 27 0f fa 00 75 d2 ac 45 8b a7 72 f2 90 a9 83 50
#> [6385] 38 41 ea 20 68 0e 95 e4 23 0c f0 40 20 fc c8 27 e3 81 f7 40 6e 38 78 47
#> [6409] 88 9e eb ad 49 1c 6a a6 c7 ae e3 39 90 33 57 3f b8 63 4a 49 6b ad 4a 8c
#> [6433] d1 da 6c 4a 4c bc a1 c9 3e 25 8a 62 42 cc 36 28 a3 8a 32 e9 d8 eb 64 f2
#> [6457] b8 8c e3 28 6d 54 f5 4d 36 4c 08 25 04 63 94 09 42 27 45 9f 10 da 77 bd
#> [6481] eb 35 5b 46 29 aa de 7b af 45 9d 4b 28 a5 42 0c 4a 30 a7 db 19 bc c7 0b
#> [6505] 65 c4 27 46 e9 7e 1f 6e 7a a8 f6 f2 2d fb f6 f3 fb 34 75 b4 7a 77 a5 1f
#> [6529] c7 79 dc c9 5d f1 7e 38 9f ff c0 dd 68 40 ea af af e4 8f 01 8f f0 04 8f
#> [6553] 4f eb 17 54 97 ed 6d 25 74 f8 fb fc bc 6d db 7a b9 ac db ba be c1 bf 9d
#> [6577] 1a 37 9c e0 f7 fc f9 e3 7c fa 0f b7 1f 15 b9 af 39 67 16 29 73 51 49 07
#> [6601] ae 3c 2a 41 3c 24 9b 4a 89 d2 b2 51 6a 80 64 a3 a9 f7 48 ed b8 2c cb cb
#> [6625] cb b2 48 49 be 99 94 52 cc 79 04 b9 80 e4 78 ac 27 2a 58 49 1f b0 61 b3
#> [6649] a5 33 cb fd 2f c0 5a 69 33 58 c8 25 11 77 3d 93 54 21 28 ca 10 3e 66 a4
#> [6673] af 5a 14 50 04 59 01 7d 54 a8 27 d9 d6 95 d1 31 41 e0 89 eb 86 5f d3 c3
#> [6697] 74 70 e0 31 fb 36 cf b7 99 83 eb 2c ae 62 a6 b8 b1 83 f8 dc c3 dd 68 d0
#> [6721] b8 0b 9a 06 8d a6 41 a3 69 d0 68 1a 34 9a 06 8d 77 76 e9 40 00 00 00 00
#> [6745] 00 c8 ff b5 11 1a 6c d0 06 6d d0 06 6d d0 06 6d d0 06 6d d0 06 6d d0 06
#> [6769] 6d d0 06 6d 80 3d 7b 59 49 28 8a c2 38 be 8e 99 a6 dd bc 64 d9 1d 2b ab
#> [6793] fd 51 9c 81 07 54 50 0c 45 70 24 42 39 d4 41 41 e0 2b 34 69 d8 33 f4 b6
#> [6817] 25 14 61 9f dd 94 c0 bd 5d bf 67 f8 e0 ec b3 fe 4a 67 a0 74 06 ea 9f 66
#> [6841] 80 eb 8a 51 f3 c3 bf 2b f3 0c 86 8a 57 05 a3 e6 44 b7 0d d0 0c de d4 fa
#> [6865] be 51 73 a0 32 c0 08 c1 a8 87 ba 51 ae 2b 35 03 f0 0c 46 75 aa 46 b9 cc
#> [6889] 7f aa e1 93 e8 fb 0c f4 89 30 2f 5a 3d 90 9c 24 c1 2e 6e 8c 72 53 f5 1a
#> [6913] 2c 93 12 89 c4 c0 6e 1b 46 b9 a7 70 55 04 3b 49 c9 ab 8d 73 b0 e0 be 64
#> [6937] 94 5b fc fe 33 d8 ca 91 bc c9 1e 82 95 ef f4 e7 d1 29 f5 07 b0 d8 b2 27
#> [6961] 1f 56 17 c0 da 5d a3 5c 51 ed 00 2c bf 23 23 e2 e9 30 d8 40 ef cb 6e f0
#> [6985] 2f cb 60 7b 6b 42 12 67 60 41 53 9f 08 0e e8 5e 80 ed 6e c9 58 c7 19 b0
#> [7009] da 93 51 76 6b 3c 82 85 93 21 f9 82 17 59 04 eb b5 8c b2 57 e9 3e 00 db
#> [7033] 4f c8 37 52 b9 28 de 69 82 76 c0 30 27 b3 c3 ac fc 60 e9 04 7a 5f 76 46
#> [7057] ab 0d b6 b0 ea 09 21 47 2b 60 cf 9a a0 ed 53 19 80 45 d3 71 f9 15 6f 39
#> [7081] 86 57 9a a0 ed 56 68 16 c1 ce d6 e5 d7 76 f2 00 34 41 db cc ef d7 c0 32
#> [7105] a7 f2 27 6b 7b 60 e5 4b fd 32 58 a2 de 03 5b 8c 78 f2 57 5b bb 9a a0 6d
#> [7129] 35 cc c9 2c 97 92 09 84 92 61 b0 47 4d d0 b3 6e 7c 4e 3e 58 92 09 25 f6
#> [7153] 35 41 5b c7 bf f9 22 27 4f 21 bb ad 09 da 2e 8d 5b b0 d8 66 48 a6 e2 45
#> [7177] c6 26 68 bd 2f cf a6 52 27 00 cb 6f c8 d4 e2 e9 a8 26 e8 17 f6 ce 66 a5
#> [7201] ad 28 0a a3 b7 45 21 28 1a b4 d6 3f a4 0d 15 ed 0f 96 0c 2c d4 41 44 b0
#> [7225] 08 0e 0b c6 a1 93 fa 30 7d ed de 4e 5a 8b 7b d7 e4 78 cf cd fe 59 eb 29
#> [7249] 0e 67 b1 bf e5 83 53 4d 27 77 c2 50 56 d0 fc 2f 1b 43 d6 c9 6b 4d 67 1c
#> [7273] cb 0a 9a 27 82 21 54 9d dc 15 ba 82 e6 7f d9 0a 67 57 8a 4e 2e 44 57 d0
#> [7297] 9f 04 6e f8 5f 36 c1 bd f4 73 bc b7 db 54 60 f0 06 05 6d 93 eb 3b 49 27
#> [7321] b7 3f c7 15 50 15 f4 94 27 42 09 46 74 72 09 2f b7 39 71 32 86 72 88 34
#> [7345] 6c aa b2 25 29 e8 2f df f9 5f 9e 07 6b 3a 59 a0 54 41 f3 44 e8 17 fd 10
#> [7369] 69 b9 3d 44 aa 8c 7e e2 74 c2 89 53 9f e8 87 48 a3 fd a6 00 14 b4 53 ba
#> [7393] d4 c9 e5 0a 9a 13 a7 85 32 9e 9e 48 3a 79 a7 e9 0f 5d 41 5f dc 7f 04 15
#> [7417] f3 3a b9 50 41 73 e2 b4 18 ce c5 43 a4 0f 45 3a 19 05 ed 14 e5 10 69 bd
#> [7441] 59 18 43 4e 9c 7a e7 56 3e 44 6a 7a 44 57 d0 ac 6c f6 c3 44 d3 c9 0b e6
#> [7465] c5 ea 32 27 4e ff c7 9f 4e 2e 61 7f c4 89 53 3f fc 14 75 f2 61 63 04 14
#> [7489] 74 1f 5c 6b 87 48 76 d8 91 14 f4 09 0a ba 3b 26 37 92 4e 6e 0f 91 4c 81
#> [7513] 82 16 f0 ae 93 0b d0 56 36 79 22 54 db b5 3c 6e 4c b2 2e af 6c f2 44 f0
#> [7537] ac 93 67 86 95 4d 81 38 3a b9 43 05 cd ff 72 f9 ae a5 f4 28 38 18 34 c6
#> [7561] 41 41 77 c9 54 d9 b5 74 c0 e1 1e 2b 9b 15 75 f2 d2 f6 c2 7f 8e ff 85 95
#> [7585] cd aa 9c 99 d2 c9 1a ac 6c ca d4 de b5 74 05 0a 3a a0 4e 2e e1 98 95 cd
#> [7609] 6e 0f 91 96 da 43 24 87 a0 a0 3b d5 c9 ef 0d fe 1c a3 a0 ab 31 16 75 f2
#> [7633] 6b 33 3a b9 84 c1 81 f4 44 60 65 53 e7 d2 be 4e 6e 61 65 b3 2a e7 2e 74
#> [7657] 72 99 82 5e 42 41 fb 39 44 aa 03 a1 c7 e7 67 12 c3 40 e8 f1 69 2e c5 43
#> [7681] 24 c3 3a 19 05 2d 50 27 93 68 5a 27 cf 06 a1 c7 e7 67 12 03 32 24 f4 a8
#> [7705] 70 eb 57 27 97 70 c8 ca a6 c0 37 a3 87 48 02 ac 6c 56 a2 d5 c9 76 0f 91
#> [7729] ea d1 2a 68 56 36 1f 20 ea e4 3d 67 3a 99 13 a7 1a 99 44 f7 3f c7 2d 84
#> [7753] 1e 0d ef 5a da 42 0e 3d fe c8 f5 44 68 33 89 a1 74 f2 5f 08 3d 3a d8 b5
#> [7777] b4 45 ee d0 e3 e5 9d b3 43 a4 6a b4 0a 3a eb ca a6 9c 49 8c a0 93 59 d9
#> [7801] 44 27 37 84 1e e7 d9 b5 0c ae 93 09 3d 7a da b5 b4 45 aa d0 a3 fc 73 7c
#> [7825] e4 e8 10 e9 11 ac 6c da ca 24 fa 47 54 d0 5f a3 85 1e c3 1c 22 3d 86 95
#> [7849] 4d 74 72 97 bc 58 0d ad a0 2d ef 5a da 62 73 14 37 f4 78 7f 91 54 27 a3
#> [7873] a0 1f ee 5a 26 d6 c9 25 bc 0b b8 b2 39 b9 c9 ad 93 59 d9 74 b5 6b 69 8b
#> [7897] ad a3 40 ff cb e8 64 42 8f 0e 77 2d 6d 11 e3 c4 c9 e3 ae a5 2d 02 28 68
#> [7921] 74 32 2b 9b ae 77 2d 6d e1 39 f4 88 4e fe 4d f2 d0 a3 ad 4c a2 7f 56 5e
#> [7945] 39 5c d9 1c 7f b6 96 49 f4 8f bf 95 4d 74 f2 1f f2 2a 68 79 d7 72 83 9f
#> [7969] e3 4c 2b 9b e8 e4 c7 e4 5b d9 34 9e 49 f4 8f 07 05 ed 20 93 e8 1f eb 2b
#> [7993] 9b 3e 32 89 fe 31 ad a0 15 9d cc cf f1 03 c2 87 1e e3 ef 5a da 62 f7 ad
#> [8017] c5 2b 68 74 f2 8c 44 56 d0 e8 e4 45 b0 b2 61 6a 65 33 d3 ae a5 2d 2c 29
#> [8041] 68 74 72 11 b1 42 8f be 33 89 fe 31 11 7a f4 9f 49 f4 cf c2 43 8f 31 32
#> [8065] 89 fe 19 cc af a0 d1 c9 11 99 7f 65 93 4c 62 44 e6 55 d0 64 12 83 32 87
#> [8089] 82 26 93 18 99 99 56 36 d1 c9 f1 99 4d 41 93 49 8c ce 13 a1 47 0e 91 b2
#> [8113] d0 4b e8 71 3c 45 27 5b 47 0d 3d 92 49 cc 85 a2 a0 c9 24 26 43 5b d9 44
#> [8137] 27 27 43 0b 3d 92 49 4c 86 16 7a 64 d7 32 17 9a 82 26 93 98 0c 2d f4 48
#> [8161] 26 31 19 da 89 13 99 c4 64 c8 0a 7a 3a 26 93 98 0b 2d f4 48 26 31 19 5a
#> [8185] e8 91 5d cb 64 68 2b 9b 1c 22 25 63 4d 0e 3d 16 66 12 47 e8 64 a7 68 2b
#> [8209] 9b e8 e4 64 68 a1 c7 79 33 89 fc 1c 3b 47 09 3d b2 6b 99 0c 2d f4 28 ea
#> [8233] 64 32 89 71 99 51 41 9f b2 6b 19 1c 2d f4 c8 21 52 32 9e 5a d9 9c a0 93
#> [8257] 53 a0 9d 38 b1 6b 99 8c 4d e5 c4 89 5d cb 64 c8 0a fa 8a 4c e2 af f6 ee
#> [8281] e0 04 40 20 86 a2 a0 36 e0 55 14 6f 62 ff 2d 1a 57 10 64 b7 83 3f 53 46
#> [8305] 1e 49 d2 ac 83 04 2d 27 e7 e9 af 6c 7a 93 18 e9 ff e8 51 4e 8e b5 7c 2b
#> [8329] 4e 72 72 b4 77 c5 c9 5d cb 74 4f 82 96 93 a9 04 6d 11 89 72 6e 72 32 2d
#> [8353] 41 bb 6b 49 4b d0 ee 5a 52 09 fa aa 71 d2 21 27 c7 9b 77 83 e3 e9 06 67
#> [8377] df ad 29 27 e5 df 6f 00 00 00 00 49 45 4e 44 ae 42 60 82