Receive Async

Receive serialised data asynchronously over a Socket (with ability to set a timeout). For sending and receiving within R.

recv_aio(socket, n = 1L, timeout, keep.raw = TRUE)

Arguments

socket

a Socket.

n

[default 1L] number of messages to receive asynchronously.

timeout

in ms. If unspecified, a socket-specific default timeout will be used.

keep.raw

[default TRUE] TRUE to keep the received raw vector (useful for verification e.g. via hashing). If FALSE, will return the unserialised object only.

Value

Named list of 2 elements: 'raw' containing a list of received raw vectors and 'data' containing a list of unserialised R objects, or else a list of unserialised R objects if keep.raw is set to FALSE. Note: a list of lists is always returned even when n = 1. To access the first raw element, for example, use $raw[[1]] and the first data element use $data[[1]].

Details

Async recv will block while awaiting all 'n' messages to arrive. Set a timeout to ensure that the function returns under all conditions.

In case of an error in unserialisation (e.g. the data was not sent serialised), the function will still return a list of received raw vectors to allow the data to be recovered.

Examples

s1 <- socket("pair", listen = "inproc://nanonext")
#> listener started...
s2 <- socket("pair", dial = "inproc://nanonext")
#> dialer started...

send_aio(s1, data.frame(a = 1, b = 2), data.frame(c = 3, d = 4), timeout = 100)
res <- recv_aio(s2, 2L, timeout = 100)
res
#> $raw
#> $raw[[1]]
#>   [1] 58 0a 00 00 00 03 00 04 01 02 00 03 05 00 00 00 00 05 55 54 46 2d 38 00 00
#>  [26] 03 13 00 00 00 02 00 00 00 0e 00 00 00 01 3f f0 00 00 00 00 00 00 00 00 00
#>  [51] 0e 00 00 00 01 40 00 00 00 00 00 00 00 00 00 04 02 00 00 00 01 00 04 00 09
#>  [76] 00 00 00 05 6e 61 6d 65 73 00 00 00 10 00 00 00 02 00 04 00 09 00 00 00 01
#> [101] 61 00 04 00 09 00 00 00 01 62 00 00 04 02 00 00 00 01 00 04 00 09 00 00 00
#> [126] 05 63 6c 61 73 73 00 00 00 10 00 00 00 01 00 04 00 09 00 00 00 0a 64 61 74
#> [151] 61 2e 66 72 61 6d 65 00 00 04 02 00 00 00 01 00 04 00 09 00 00 00 09 72 6f
#> [176] 77 2e 6e 61 6d 65 73 00 00 00 0d 00 00 00 02 80 00 00 00 ff ff ff ff 00 00
#> [201] 00 fe
#> 
#> $raw[[2]]
#>   [1] 58 0a 00 00 00 03 00 04 01 02 00 03 05 00 00 00 00 05 55 54 46 2d 38 00 00
#>  [26] 03 13 00 00 00 02 00 00 00 0e 00 00 00 01 40 08 00 00 00 00 00 00 00 00 00
#>  [51] 0e 00 00 00 01 40 10 00 00 00 00 00 00 00 00 04 02 00 00 00 01 00 04 00 09
#>  [76] 00 00 00 05 6e 61 6d 65 73 00 00 00 10 00 00 00 02 00 04 00 09 00 00 00 01
#> [101] 63 00 04 00 09 00 00 00 01 64 00 00 04 02 00 00 00 01 00 04 00 09 00 00 00
#> [126] 05 63 6c 61 73 73 00 00 00 10 00 00 00 01 00 04 00 09 00 00 00 0a 64 61 74
#> [151] 61 2e 66 72 61 6d 65 00 00 04 02 00 00 00 01 00 04 00 09 00 00 00 09 72 6f
#> [176] 77 2e 6e 61 6d 65 73 00 00 00 0d 00 00 00 02 80 00 00 00 ff ff ff ff 00 00
#> [201] 00 fe
#> 
#> 
#> $data
#> $data[[1]]
#>   a b
#> 1 1 2
#> 
#> $data[[2]]
#>   c d
#> 1 3 4
#> 
#> 

send_aio(s1, data.frame(a = 1, b = 2), data.frame(c = 3, d = 4), timeout = 100)
recv_aio(s2, 2L, timeout = 100, keep.raw = FALSE)
#> [[1]]
#>   a b
#> 1 1 2
#> 
#> [[2]]
#>   c d
#> 1 3 4
#> 

close(s1)
#> socket closed
close(s2)
#> socket closed