Set and Retrieve Refhook Function

Sets the refhook function stored internally, used in all message send and receive operations. Calling without an argument retrieves the stored value.

refhook(x = list())

Arguments

x

a function or function encoded by base64enc, or NULL to reset.

Value

The current refhook function stored internally.

Details

The provided function maps to the 'refhook' argument of serialize or unserialize, as the case may be. As the same function is used in both cases, it should contain the logic to distinguish between them as necessary.

Examples

refhook()
#> NULL

func <- function(x) if (is_nano(x)) base64enc(x) else
    if (is.character(x)) unserialize(base64dec(x, convert = FALSE))

refhook(func)
#> function(x) if (is_nano(x)) base64enc(x) else
#>     if (is.character(x)) unserialize(base64dec(x, convert = FALSE))
#> <environment: 0x562feae26ab8>

enc <- base64enc(func)
enc
#> [1] "WAoAAAADAAQDAQADBQAAAAAFVVRGLTgAAAYDAAAEAgAAAAEABAAJAAAABnNyY3JlZgAAAw0AAAAIAAAAAwAAAAkAAAAEAAAAQwAAAAkAAABDAAAAAwAAAAQAAAQCAAAAAQAEAAkAAAAHc3JjZmlsZQAAAAQAAAAAAAAA8gAAAP4AAAATAAAAHQAAAP4AAAD+AAAA/gAABAIAAAABAAQACQAAAAJ3ZAAAABAAAAABAAQACQAAACkvdG1wL1J0bXBaOWdJYjcvZmlsZTFjYzExOTQ1MzIwL3JlZmVyZW5jZQAAAP4AAAD+AAAEAgAAAAEABAAJAAAABWxpbmVzAAAAEAAAABAABAAJAAAACXJlZmhvb2soKQAEAAkAAAAAAAQACQAAADVmdW5jIDwtIGZ1bmN0aW9uKHgpIGlmIChpc19uYW5vKHgpKSBiYXNlNjRlbmMoeCkgZWxzZQAEAAkAAABDICAgIGlmIChpcy5jaGFyYWN0ZXIoeCkpIHVuc2VyaWFsaXplKGJhc2U2NGRlYyh4LCBjb252ZXJ0ID0gRkFMU0UpKQAEAAkAAAAAAAQACQAAAA1yZWZob29rKGZ1bmMpAAQACQAAAAAABAAJAAAAFmVuYyA8LSBiYXNlNjRlbmMoZnVuYykABAAJAAAAA2VuYwAEAAkAAAAMcmVmaG9vayhlbmMpAAQACQAAAAAABAAJAAAACXJlZmhvb2soKQAEAAkAAAAAAAQACQAAAA1yZWZob29rKE5VTEwpAAQACQAAAAAABAAJAAAAAAAAAP4AAAQCAAAAAQAEAAkAAAADRW5jAAAAEAAAAAEABAAJAAAAB3Vua25vd24AAAD+AAAA/gAABAIAAAABAAQACQAAAAZpc0ZpbGUAAAAKAAAAAQAAAAAAAAD+AAAEAgAAAAEABAAJAAAACXRpbWVzdGFtcAAAAw4AAAABQdkkaF03oooAAAQCAAAAAQAEAAkAAAAFY2xhc3MAAAAQAAAAAgAEAAkAAAAHUE9TSVhjdAAEAAkAAAAGUE9TSVh0AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAQCAAAAAQAEAAkAAAAIZmlsZW5hbWUAAAAQAAAAAQAEAAkAAAAGPHRleHQ+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAQCAAAAAQAEAAkAAAANZml4ZWROZXdsaW5lcwAAAAoAAAABAAAAAQAAAP4AAAQCAAAAAQAEAAkAAAAJcGFyc2VEYXRhAAADDQAAAyAAAAABAAAAAQAAAAEAAAAHAAAAAQAAASgAAAABAAAAAwAAAAEAAAAIAAAAAQAAAAgAAAABAAAAKAAAAAIAAAAGAAAAAQAAAAEAAAABAAAABwAAAAAAAABPAAAAAwAAAAYAAAABAAAACQAAAAEAAAAJAAAAAQAAACkAAAAEAAAABgAAAAEAAAABAAAAAQAAAAkAAAAAAAAATwAAAAYAAAAAAAAAAwAAAAEAAAADAAAABAAAAAEAAAEHAAAADAAAAA4AAAADAAAABgAAAAMAAAAHAAAAAQAAAQoAAAANAAAAYAAAAAMAAAABAAAAAwAAAAQAAAAAAAAATwAAAA4AAABgAAAAAwAAAAkAAAADAAAAEAAAAAEAAAEIAAAADwAAAF8AAAADAAAAEQAAAAMAAAARAAAAAQAAACgAAAAQAAAAXwAAAAMAAAASAAAAAwAAABIAAAABAAABJAAAABEAAABfAAAAAwAAABMAAAADAAAAEwAAAAEAAAApAAAAEgAAAF8AAAADAAAAFQAAAAMAAAAWAAAAAQAAARAAAAAUAAAAXQAAAAMAAAAYAAAAAwAAABgAAAABAAAAKAAAABUAAABdAAAAAwAAABkAAAADAAAAHwAAAAEAAAEoAAAAFgAAABgAAAADAAAAIAAAAAMAAAAgAAAAAQAAACgAAAAXAAAAHwAAAAMAAAAZAAAAAwAAAB8AAAAAAAAATwAAABgAAAAfAAAAAwAAACEAAAADAAAAIQAAAAEAAAEHAAAAGQAAABsAAAADAAAAIgAAAAMAAAAiAAAAAQAAACkAAAAaAAAAHwAAAAMAAAAhAAAAAwAAACEAAAAAAAAATwAAABsAAAAfAAAAAwAAABkAAAADAAAAIgAAAAAAAABPAAAAHwAAAF0AAAADAAAAIwAAAAMAAAAjAAAAAQAAACkAAAAgAAAAXQAAAAMAAAAlAAAAAwAAAC0AAAABAAABKAAAACMAAAAlAAAAAwAAAC4AAAADAAAALgAAAAEAAAAoAAAAJAAAACwAAAADAAAAJQAAAAMAAAAtAAAAAAAAAE8AAAAlAAAALAAAAAMAAAAvAAAAAwAAAC8AAAABAAABBwAAACYAAAAoAAAAAwAAADAAAAADAAAAMAAAAAEAAAApAAAAJwAAACwAAAADAAAALwAAAAMAAAAvAAAAAAAAAE8AAAAoAAAALAAAAAMAAAAlAAAAAwAAADAAAAAAAAAATwAAACwAAABdAAAAAwAAADIAAAADAAAANQAAAAEAAAERAAAALQAAAF0AAAAEAAAABQAAAAQAAAAGAAAAAQAAARAAAAAwAAAAWwAAAAQAAAAIAAAABAAAAAgAAAABAAAAKAAAADEAAABbAAAABAAAAAkAAAAEAAAAFAAAAAEAAAEoAAAAMgAAADQAAAAEAAAAFQAAAAQAAAAVAAAAAQAAACgAAAAzAAAAOwAAAAQAAAAJAAAABAAAABQAAAAAAAAATwAAADQAAAA7AAAABAAAABYAAAAEAAAAFgAAAAEAAAEHAAAANQAAADcAAAAEAAAAFwAAAAQAAAAXAAAAAQAAACkAAAA2AAAAOwAAAAQAAAAWAAAABAAAABYAAAAAAAAATwAAADcAAAA7AAAABAAAAAkAAAAEAAAAFwAAAAAAAABPAAAAOwAAAFsAAAAEAAAAGAAAAAQAAAAYAAAAAQAAACkAAAA8AAAAWwAAAAQAAAAaAAAABAAAACQAAAABAAABKAAAAD8AAABBAAAABAAAACUAAAAEAAAAJQAAAAEAAAAoAAAAQAAAAFgAAAAEAAAAGgAAAAQAAAAkAAAAAAAAAE8AAABBAAAAWAAAAAQAAAAmAAAABAAAAC4AAAABAAABKAAAAEIAAABEAAAABAAAAC8AAAAEAAAALwAAAAEAAAAoAAAAQwAAAFMAAAAEAAAAJgAAAAQAAAAuAAAAAAAAAE8AAABEAAAAUwAAAAQAAAAwAAAABAAAADAAAAABAAABBwAAAEUAAABHAAAABAAAADEAAAAEAAAAMQAAAAEAAAAsAAAARgAAAFMAAAAEAAAAMAAAAAQAAAAwAAAAAAAAAE8AAABHAAAAUwAAAAQAAAAzAAAABAAAADkAAAABAAABJwAAAEsAAABTAAAABAAAADsAAAAEAAAAOwAAAAEAAAEmAAAATAAAAFMAAAAEAAAAPQAAAAQAAABBAAAAAQAAAQUAAABNAAAATgAAAAQAAAA9AAAABAAAAEEAAAAAAAAATwAAAE4AAABTAAAABAAAAEIAAAAEAAAAQgAAAAEAAAApAAAATwAAAFMAAAAEAAAAJgAAAAQAAABCAAAAAAAAAE8AAABTAAAAWAAAAAQAAABDAAAABAAAAEMAAAABAAAAKQAAAFQAAABYAAAABAAAABoAAAAEAAAAQwAAAAAAAABPAAAAWAAAAFsAAAAEAAAABQAAAAQAAABDAAAAAAAAAE8AAABbAAAAXQAAAAMAAAAVAAAABAAAAEMAAAAAAAAATwAAAF0AAABfAAAAAwAAAAkAAAAEAAAAQwAAAAAAAABPAAAAXwAAAGAAAAADAAAAAQAAAAQAAABDAAAAAAAAAE8AAABgAAAAAAAAAAYAAAABAAAABgAAAAcAAAABAAABKAAAAGUAAABnAAAABgAAAAgAAAAGAAAACAAAAAEAAAAoAAAAZgAAAG4AAAAGAAAAAQAAAAYAAAAHAAAAAAAAAE8AAABnAAAAbgAAAAYAAAAJAAAABgAAAAwAAAABAAABBwAAAGgAAABqAAAABgAAAA0AAAAGAAAADQAAAAEAAAApAAAAaQAAAG4AAAAGAAAACQAAAAYAAAAMAAAAAAAAAE8AAABqAAAAbgAAAAYAAAABAAAABgAAAA0AAAAAAAAATwAAAG4AAAAAAAAACAAAAAEAAAAIAAAAAwAAAAEAAAEHAAAAdAAAAHYAAAAIAAAABQAAAAgAAAAGAAAAAQAAAQoAAAB1AAAAggAAAAgAAAABAAAACAAAAAMAAAAAAAAATwAAAHYAAACCAAAACAAAAAgAAAAIAAAAEAAAAAEAAAEoAAAAdwAAAHkAAAAIAAAAEQAAAAgAAAARAAAAAQAAACgAAAB4AAAAgAAAAAgAAAAIAAAACAAAABAAAAAAAAAATwAAAHkAAACAAAAACAAAABIAAAAIAAAAFQAAAAEAAAEHAAAAegAAAHwAAAAIAAAAFgAAAAgAAAAWAAAAAQAAACkAAAB7AAAAgAAAAAgAAAASAAAACAAAABUAAAAAAAAATwAAAHwAAACAAAAACAAAAAgAAAAIAAAAFgAAAAAAAABPAAAAgAAAAIIAAAAIAAAAAQAAAAgAAAAWAAAAAAAAAE8AAACCAAAAAAAAAAkAAAABAAAACQAAAAMAAAABAAABBwAAAIUAAACHAAAACQAAAAEAAAAJAAAAAwAAAAAAAABPAAAAhwAAAAAAAAAKAAAAAQAAAAoAAAAHAAAAAQAAASgAAACKAAAAjAAAAAoAAAAIAAAACgAAAAgAAAABAAAAKAAAAIsAAACTAAAACgAAAAEAAAAKAAAABwAAAAAAAABPAAAAjAAAAJMAAAAKAAAACQAAAAoAAAALAAAAAQAAAQcAAACNAAAAjwAAAAoAAAAMAAAACgAAAAwAAAABAAAAKQAAAI4AAACTAAAACgAAAAkAAAAKAAAACwAAAAAAAABPAAAAjwAAAJMAAAAKAAAAAQAAAAoAAAAMAAAAAAAAAE8AAACTAAAAAAAAAAwAAAABAAAADAAAAAcAAAABAAABKAAAAJkAAACbAAAADAAAAAgAAAAMAAAACAAAAAEAAAAoAAAAmgAAAJ4AAAAMAAAAAQAAAAwAAAAHAAAAAAAAAE8AAACbAAAAngAAAAwAAAAJAAAADAAAAAkAAAABAAAAKQAAAJwAAACeAAAADAAAAAEAAAAMAAAACQAAAAAAAABPAAAAngAAAAAAAAAOAAAAAQAAAA4AAAAHAAAAAQAAASgAAACkAAAApgAAAA4AAAAIAAAADgAAAAgAAAABAAAAKAAAAKUAAACtAAAADgAAAAEAAAAOAAAABwAAAAAAAABPAAAApgAAAK0AAAAOAAAACQAAAA4AAAAMAAAAAQAAAQYAAACnAAAAqQAAAA4AAAANAAAADgAAAA0AAAABAAAAKQAAAKgAAACtAAAADgAAAAkAAAAOAAAADAAAAAAAAABPAAAAqQAAAK0AAAAOAAAAAQAAAA4AAAANAAAAAAAAAE8AAACtAAAAAAAABAIAAAABAAQACQAAAANkaW0AAAANAAAAAgAAAAgAAABkAAAEAgAAAAEABAAJAAAABnRva2VucwAAABAAAABkAAQACQAAABRTWU1CT0xfRlVOQ1RJT05fQ0FMTAAEAAkAAAADJygnAAQACQAAAARleHByAAQACQAAAAMnKScABAAJAAAABGV4cHIABAAJAAAABlNZTUJPTAAEAAkAAAALTEVGVF9BU1NJR04ABAAJAAAABGV4cHIABAAJAAAACEZVTkNUSU9OAAQACQAAAAMnKCcABAAJAAAADlNZTUJPTF9GT1JNQUxTAAQACQAAAAMnKScABAAJAAAAAklGAAQACQAAAAMnKCcABAAJAAAAFFNZTUJPTF9GVU5DVElPTl9DQUxMAAQACQAAAAMnKCcABAAJAAAABGV4cHIABAAJAAAABlNZTUJPTAAEAAkAAAADJyknAAQACQAAAARleHByAAQACQAAAARleHByAAQACQAAAAMnKScABAAJAAAAFFNZTUJPTF9GVU5DVElPTl9DQUxMAAQACQAAAAMnKCcABAAJAAAABGV4cHIABAAJAAAABlNZTUJPTAAEAAkAAAADJyknAAQACQAAAARleHByAAQACQAAAARleHByAAQACQAAAARFTFNFAAQACQAAAAJJRgAEAAkAAAADJygnAAQACQAAABRTWU1CT0xfRlVOQ1RJT05fQ0FMTAAEAAkAAAADJygnAAQACQAAAARleHByAAQACQAAAAZTWU1CT0wABAAJAAAAAycpJwAEAAkAAAAEZXhwcgAEAAkAAAAEZXhwcgAEAAkAAAADJyknAAQACQAAABRTWU1CT0xfRlVOQ1RJT05fQ0FMTAAEAAkAAAADJygnAAQACQAAAARleHByAAQACQAAABRTWU1CT0xfRlVOQ1RJT05fQ0FMTAAEAAkAAAADJygnAAQACQAAAARleHByAAQACQAAAAZTWU1CT0wABAAJAAAAAycsJwAEAAkAAAAEZXhwcgAEAAkAAAAKU1lNQk9MX1NVQgAEAAkAAAAGRVFfU1VCAAQACQAAAAlOVU1fQ09OU1QABAAJAAAABGV4cHIABAAJAAAAAycpJwAEAAkAAAAEZXhwcgAEAAkAAAADJyknAAQACQAAAARleHByAAQACQAAAARleHByAAQACQAAAARleHByAAQACQAAAARleHByAAQACQAAAARleHByAAQACQAAABRTWU1CT0xfRlVOQ1RJT05fQ0FMTAAEAAkAAAADJygnAAQACQAAAARleHByAAQACQAAAAZTWU1CT0wABAAJAAAAAycpJwAEAAkAAAAEZXhwcgAEAAkAAAAEZXhwcgAEAAkAAAAGU1lNQk9MAAQACQAAAAtMRUZUX0FTU0lHTgAEAAkAAAAEZXhwcgAEAAkAAAAUU1lNQk9MX0ZVTkNUSU9OX0NBTEwABAAJAAAAAycoJwAEAAkAAAAEZXhwcgAEAAkAAAAGU1lNQk9MAAQACQAAAAMnKScABAAJAAAABGV4cHIABAAJAAAABGV4cHIABAAJAAAABGV4cHIABAAJAAAABlNZTUJPTAAEAAkAAAAEZXhwcgAEAAkAAAAUU1lNQk9MX0ZVTkNUSU9OX0NBTEwABAAJAAAAAycoJwAEAAkAAAAEZXhwcgAEAAkAAAAGU1lNQk9MAAQACQAAAAMnKScABAAJAAAABGV4cHIABAAJAAAABGV4cHIABAAJAAAAFFNZTUJPTF9GVU5DVElPTl9DQUxMAAQACQAAAAMnKCcABAAJAAAABGV4cHIABAAJAAAAAycpJwAEAAkAAAAEZXhwcgAEAAkAAAAUU1lNQk9MX0ZVTkNUSU9OX0NBTEwABAAJAAAAAycoJwAEAAkAAAAEZXhwcgAEAAkAAAAKTlVMTF9DT05TVAAEAAkAAAADJyknAAQACQAAAARleHByAAQACQAAAARleHByAAAEAgAAAAEABAAJAAAABHRleHQAAAAQAAAAZAAEAAkAAAAHcmVmaG9vawAEAAkAAAABKAAEAAkAAAAAAAQACQAAAAEpAAQACQAAAAAABAAJAAAABGZ1bmMABAAJAAAAAjwtAAQACQAAAAAABAAJAAAACGZ1bmN0aW9uAAQACQAAAAEoAAQACQAAAAF4AAQACQAAAAEpAAQACQAAAAJpZgAEAAkAAAABKAAEAAkAAAAHaXNfbmFubwAEAAkAAAABKAAEAAkAAAAAAAQACQAAAAF4AAQACQAAAAEpAAQACQAAAAAABAAJAAAAAAAEAAkAAAABKQAEAAkAAAAJYmFzZTY0ZW5jAAQACQAAAAEoAAQACQAAAAAABAAJAAAAAXgABAAJAAAAASkABAAJAAAAAAAEAAkAAAAAAAQACQAAAARlbHNlAAQACQAAAAJpZgAEAAkAAAABKAAEAAkAAAAMaXMuY2hhcmFjdGVyAAQACQAAAAEoAAQACQAAAAAABAAJAAAAAXgABAAJAAAAASkABAAJAAAAAAAEAAkAAAAAAAQACQAAAAEpAAQACQAAAAt1bnNlcmlhbGl6ZQAEAAkAAAABKAAEAAkAAAAAAAQACQAAAAliYXNlNjRkZWMABAAJAAAAASgABAAJAAAAAAAEAAkAAAABeAAEAAkAAAABLAAEAAkAAAAAAAQACQAAAAdjb252ZXJ0AAQACQAAAAE9AAQACQAAAAVGQUxTRQAEAAkAAAAAAAQACQAAAAEpAAQACQAAAAAABAAJAAAAASkABAAJAAAAAAAEAAkAAAAAAAQACQAAAAAABAAJAAAAAAAEAAkAAAAAAAQACQAAAAdyZWZob29rAAQACQAAAAEoAAQACQAAAAAABAAJAAAABGZ1bmMABAAJAAAAASkABAAJAAAAAAAEAAkAAAAAAAQACQAAAANlbmMABAAJAAAAAjwtAAQACQAAAAAABAAJAAAACWJhc2U2NGVuYwAEAAkAAAABKAAEAAkAAAAAAAQACQAAAARmdW5jAAQACQAAAAEpAAQACQAAAAAABAAJAAAAAAAEAAkAAAAAAAQACQAAAANlbmMABAAJAAAAAAAEAAkAAAAHcmVmaG9vawAEAAkAAAABKAAEAAkAAAAAAAQACQAAAANlbmMABAAJAAAAASkABAAJAAAAAAAEAAkAAAAAAAQACQAAAAdyZWZob29rAAQACQAAAAEoAAQACQAAAAAABAAJAAAAASkABAAJAAAAAAAEAAkAAAAHcmVmaG9vawAEAAkAAAABKAAEAAkAAAAAAAQACQAAAAROVUxMAAQACQAAAAEpAAQACQAAAAAABAAJAAAAAAAABAIAAAn/AAAAEAAAAAEABAAJAAAACXBhcnNlRGF0YQAAAP4AAAD+AAAEAgAACf8AAAAQAAAAAgAEAAkAAAALc3JjZmlsZWNvcHkABAAJAAAAB3NyY2ZpbGUAAAD+AAAEAgAACf8AAAAQAAAAAQAEAAkAAAAGc3JjcmVmAAAA/gAAAP4AAAAEAAAAAAAAAAQAAAAAAAAABAAAAAAAAAAEAAAAAAAAAP0AAAD+AAAAEwAAAB0AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAAEwAAAB0AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAAEwAAAB0AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAAEwAAAB0AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAEAgAAAAEABAAJAAAABGZ1bmMAAAYDAAAEAgAAAf8AAAMNAAAACAAAAAMAAAAJAAAABAAAAEMAAAAJAAAAQwAAAAMAAAAEAAAEAgAAAv8AAAP/AAAEAgAACf8AAAAQAAAAAQAEAAkAAAAGc3JjcmVmAAAA/gAAAP4AABD/AAAEAgAAAAEABAAJAAAAAXgAAAD7AAAA/gAAAAYAAAABAAQACQAAAAJpZgAAAAIAAAAGAAAAAQAEAAkAAAAHaXNfbmFubwAAAAIAABX/AAAA/gAAAAIAAAAGAAAAAQAEAAkAAAAJYmFzZTY0ZW5jAAAAAgAAFf8AAAD+AAAAAgAAAAYAABb/AAAAAgAAAAYAAAABAAQACQAAAAxpcy5jaGFyYWN0ZXIAAAACAAAV/wAAAP4AAAACAAAABgAAAAEABAAJAAAAC3Vuc2VyaWFsaXplAAAAAgAAAAYAAAABAAQACQAAAAliYXNlNjRkZWMAAAACAAAV/wAABAIAAAABAAQACQAAAAdjb252ZXJ0AAAACgAAAAEAAAAAAAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAEAgAAFf8AAAD7AAAA/gAAAAYAABb/AAAAAgAAAAYAABf/AAAAAgAAFf8AAAD+AAAAAgAAAAYAABj/AAAAAgAAFf8AAAD+AAAAAgAAAAYAABb/AAAAAgAAAAYAABn/AAAAAgAAFf8AAAD+AAAAAgAAAAYAABr/AAAAAgAAAAYAABv/AAAAAgAAFf8AAAQCAAAc/wAAAAoAAAABAAAAAAAAAP4AAAD+AAAA/gAAAP4="
refhook(enc)
#> function(x) if (is_nano(x)) base64enc(x) else
#>     if (is.character(x)) unserialize(base64dec(x, convert = FALSE))
#> <environment: 0x562feb134ed0>

refhook()
#> function(x) if (is_nano(x)) base64enc(x) else
#>     if (is.character(x)) unserialize(base64dec(x, convert = FALSE))
#> <environment: 0x562feb134ed0>

refhook(NULL)
#> NULL