Cryptographic Hashing Using the SHA-2 Algorithms

Returns a SHA-256, SHA-224, SHA-384, or SHA-512 hash or HMAC of the supplied R object. Uses the optimised implementation from the Mbed TLS library.

sha256(x, key = NULL)

sha224(x, key = NULL)

sha384(x, key = NULL)

sha512(x, key = NULL)

Arguments

x

an object.

key

(optional) supply a secret key to generate an HMAC. If missing or NULL, the SHA-256/224/384/512 hash of 'x' is returned.

Value

A 'nanoHash' object - raw vector of length 32 for SHA-256, 28 for SHA-224, 48 for SHA-384, and 64 for SHA-512.

Details

For arguments 'x' and 'key', a raw vector is hashed directly, a scalar character string is translated to raw without serialisation, whilst all other objects are serialised first.

Use as.character() to convert the returned raw vector to a single character string.

Examples

# SHA-256 hash:
sha256("hello world!")
#> 75 09 e5 bd a0 c7 62 d2 ba c7 f9 0d 75 8b 5b 22 63 fa 01 cc bc 54 2a b5 e3 df 16 3b e0 8e 6c a9

# Convert to character string:
as.character(sha256("hello world!"))
#> [1] "7509e5bda0c762d2bac7f90d758b5b2263fa01ccbc542ab5e3df163be08e6ca9"

# Obtain HMAC:
sha256("hello world!", "SECRET_KEY")
#> 7f b6 77 26 fa 6d d0 51 04 b5 02 91 06 49 cb 7a e3 54 a1 d6 74 40 0d 7b 12 b5 a5 bb f3 98 26 95

# SHA-224 hash:
sha224("hello world!")
#> 42 8c 16 b4 30 9e 82 4c fa 87 4f e2 4c e2 af 28 94 fc af 8d 72 af 4a 36 8d 49 2e 34

# Convert to character string:
as.character(sha224("hello world!"))
#> [1] "428c16b4309e824cfa874fe24ce2af2894fcaf8d72af4a368d492e34"

# Obtain HMAC:
sha224("hello world!", "SECRET_KEY")
#> bb d5 5d 45 78 8e 14 57 00 07 cc b6 f5 71 8c f0 64 59 f1 75 2a 24 43 0c 17 2d 3d fb

# SHA-384 hash:
sha384("hello world!")
#> d3 3d 40 f7 01 0c e3 4a a8 6e fd 35 36 30 30 9e d5 c3 d7 ff ac 66 d9 88 82 5c f6 99 f4 80 3c cd f3 f0 33 23 06 12 f0 94 53 32 fb 58 0d 8a f8 05

# Convert to character string:
as.character(sha384("hello world!"))
#> [1] "d33d40f7010ce34aa86efd353630309ed5c3d7ffac66d988825cf699f4803ccdf3f033230612f0945332fb580d8af805"

# Obtain HMAC:
sha384("hello world!", "SECRET_KEY")
#> f4 4e 6c 74 9a 79 e9 ad ab d1 69 1f f7 de 05 f0 be e7 5b cd f5 7b f6 b0 b2 46 98 33 68 3b 5d 93 e7 0d 6c 8f 7e a3 a9 f9 80 27 b9 5d 04 e7 49 4e

# SHA-512 hash:
sha512("hello world!")
#> db 9b 1c d3 26 2d ee 37 75 6a 09 b9 06 49 73 58 98 47 ca a8 e5 3d 31 a9 d1 42 ea 27 01 b1 b2 8a bd 97 83 8b b9 a2 70 68 ba 30 5d c8 d0 4a 45 a1 fc f0 79 de 54 d6 07 66 69 96 b3 cc 54 f6 b6 7c

# Convert to character string:
as.character(sha512("hello world!"))
#> [1] "db9b1cd3262dee37756a09b9064973589847caa8e53d31a9d142ea2701b1b28abd97838bb9a27068ba305dc8d04a45a1fcf079de54d607666996b3cc54f6b67c"

# Obtain HMAC:
sha512("hello world!", "SECRET_KEY")
#> 65 f2 94 47 2e 69 85 ff 29 05 11 bc 10 41 56 82 25 61 1d 71 2c a5 7b 54 e8 b7 1f 3a c4 23 de dc 0d 5b 19 30 18 6d cf 85 90 a6 16 16 0b ac 65 84 4c d4 5c 1d 70 8c 40 2f b8 8d 9b 9c 16 71 61 e0