Cryptographic Hashing Using the SHA-384 Algorithm

Returns a SHA-384 hash or HMAC of the supplied R object. Uses the optimised implementation from the Mbed TLS library.

sha384(x, key = NULL)

Arguments

x

an object.

key

(optional) supply a secret key to generate an HMAC. If missing or NULL, the SHA-384 hash of 'x' is returned.

Value

A 'nanoHash' object - raw vector of length 48.

Details

For arguments 'x' and 'key', a raw vector is hashed directly, a character string is converted using charToRaw, whilst other objects are serialised first.

Use as.character() to convert the returned raw vector to a single character string.

Examples

sha384("hello world!")
#> d3 3d 40 f7 01 0c e3 4a a8 6e fd 35 36 30 30 9e d5 c3 d7 ff ac 66 d9 88 82 5c f6 99 f4 80 3c cd f3 f0 33 23 06 12 f0 94 53 32 fb 58 0d 8a f8 05

# Converts to a character string:
as.character(sha384("hello world!"))
#> [1] "d33d40f7010ce34aa86efd353630309ed5c3d7ffac66d988825cf699f4803ccdf3f033230612f0945332fb580d8af805"

# Obtain HMAC:
sha384("hello world!", "SECRET_KEY")
#> f4 4e 6c 74 9a 79 e9 ad ab d1 69 1f f7 de 05 f0 be e7 5b cd f5 7b f6 b0 b2 46 98 33 68 3b 5d 93 e7 0d 6c 8f 7e a3 a9 f9 80 27 b9 5d 04 e7 49 4e