ncurl

nano cURL - a minimalistic http(s) client.

ncurl(http)

Arguments

http

the URL/address of the resource to retrieve.

Value

Named list of 2 elements:

$raw - a raw vector of the received resource (may be saved to a file using writeBin to re-create the original)

$data - the raw vector converted to a character string (if the served content was a recognised text format), allowing further parsing within R as html, json, xml etc., or NULL otherwise (if the content was a binary file etc.)

Redirects

In interactive sessions, will prompt upon receiving a redirect location whether to follow or not (default: yes). In non-interactive sessions, redirects are never followed.

TLS Support

Connecting to secure https sites is supported if your version of the NNG library was built with TLS support (using Mbed TLS) and the environment variable 'NANONEXT_TLS' was set when installing the package e.g. by Sys.setenv(NANONEXT_TLS=1). Note: not applicable for Windows systems.

Examples

ncurl("http://httpbin.org/headers")
#> $raw
#>   [1] 7b 0a 20 20 22 68 65 61 64 65 72 73 22 3a 20 7b 0a 20 20 20 20 22 48 6f 73
#>  [26] 74 22 3a 20 22 68 74 74 70 62 69 6e 2e 6f 72 67 22 2c 20 0a 20 20 20 20 22
#>  [51] 58 2d 41 6d 7a 6e 2d 54 72 61 63 65 2d 49 64 22 3a 20 22 52 6f 6f 74 3d 31
#>  [76] 2d 36 31 66 66 63 34 39 63 2d 31 37 32 62 61 36 62 65 34 62 66 38 31 65 62
#> [101] 38 36 62 32 31 30 33 36 30 22 0a 20 20 7d 0a 7d 0a
#> 
#> $data
#> [1] "{\n  \"headers\": {\n    \"Host\": \"httpbin.org\", \n    \"X-Amzn-Trace-Id\": \"Root=1-61ffc49c-172ba6be4bf81eb86b210360\"\n  }\n}\n"
#>