ncurl

nano cURL - a minimalist http(s) client.

ncurl(http, ...)

Arguments

http

the URL address.

...

(optional) additional arguments, see 'methods' section below.

Value

Named list of 2 elements:

$raw - raw vector of the received resource (use writeBin to save to a file).

$data - converted character string (if a recognised text format), or NULL otherwise. Other tools can be used to further parse this as html, json, xml etc. if required.

Methods

Additional arguments may be passed in using '...' for HTTP methods other than GET.

Parsed as follows: [method], [content-type], [data]

Example: "POST", "text/plain", "hello world"

All 3 arguments must be supplied, and will be ignored otherwise, as will extra arguments

Redirects

In interactive sessions: will prompt upon receiving a redirect location whether to follow or not (default: yes).

In non-interactive sessions: redirects are never followed.

TLS Support

Connecting to secure https sites is supported if nng_version shows 'TLS supported'.

Examples

ncurl("http://httpbin.org/get")
#> $raw
#>   [1] 7b 0a 20 20 22 61 72 67 73 22 3a 20 7b 7d 2c 20 0a 20 20 22 68 65 61 64 65
#>  [26] 72 73 22 3a 20 7b 0a 20 20 20 20 22 48 6f 73 74 22 3a 20 22 68 74 74 70 62
#>  [51] 69 6e 2e 6f 72 67 22 2c 20 0a 20 20 20 20 22 58 2d 41 6d 7a 6e 2d 54 72 61
#>  [76] 63 65 2d 49 64 22 3a 20 22 52 6f 6f 74 3d 31 2d 36 32 32 33 63 35 65 34 2d
#> [101] 35 61 65 66 66 33 63 32 37 63 37 36 37 33 66 38 36 31 65 35 35 30 33 36 22
#> [126] 0a 20 20 7d 2c 20 0a 20 20 22 6f 72 69 67 69 6e 22 3a 20 22 31 30 34 2e 32
#> [151] 31 35 2e 31 31 33 2e 31 34 31 22 2c 20 0a 20 20 22 75 72 6c 22 3a 20 22 68
#> [176] 74 74 70 3a 2f 2f 68 74 74 70 62 69 6e 2e 6f 72 67 2f 67 65 74 22 0a 7d 0a
#> 
#> $data
#> [1] "{\n  \"args\": {}, \n  \"headers\": {\n    \"Host\": \"httpbin.org\", \n    \"X-Amzn-Trace-Id\": \"Root=1-6223c5e4-5aeff3c27c7673f861e55036\"\n  }, \n  \"origin\": \"104.215.113.141\", \n  \"url\": \"http://httpbin.org/get\"\n}\n"
#> 
ncurl("http://httpbin.org/post", "POST", "text-plain", "hello world")
#> $raw
#>   [1] 7b 0a 20 20 22 61 72 67 73 22 3a 20 7b 7d 2c 20 0a 20 20 22 64 61 74 61 22
#>  [26] 3a 20 22 68 65 6c 6c 6f 20 77 6f 72 6c 64 22 2c 20 0a 20 20 22 66 69 6c 65
#>  [51] 73 22 3a 20 7b 7d 2c 20 0a 20 20 22 66 6f 72 6d 22 3a 20 7b 7d 2c 20 0a 20
#>  [76] 20 22 68 65 61 64 65 72 73 22 3a 20 7b 0a 20 20 20 20 22 43 6f 6e 74 65 6e
#> [101] 74 2d 4c 65 6e 67 74 68 22 3a 20 22 31 31 22 2c 20 0a 20 20 20 20 22 43 6f
#> [126] 6e 74 65 6e 74 2d 54 79 70 65 22 3a 20 22 74 65 78 74 2d 70 6c 61 69 6e 22
#> [151] 2c 20 0a 20 20 20 20 22 48 6f 73 74 22 3a 20 22 68 74 74 70 62 69 6e 2e 6f
#> [176] 72 67 22 2c 20 0a 20 20 20 20 22 58 2d 41 6d 7a 6e 2d 54 72 61 63 65 2d 49
#> [201] 64 22 3a 20 22 52 6f 6f 74 3d 31 2d 36 32 32 33 63 35 65 34 2d 30 33 39 31
#> [226] 35 65 35 61 35 65 64 62 63 64 39 33 33 32 63 39 62 38 38 38 22 0a 20 20 7d
#> [251] 2c 20 0a 20 20 22 6a 73 6f 6e 22 3a 20 6e 75 6c 6c 2c 20 0a 20 20 22 6f 72
#> [276] 69 67 69 6e 22 3a 20 22 31 30 34 2e 32 31 35 2e 31 31 33 2e 31 34 31 22 2c
#> [301] 20 0a 20 20 22 75 72 6c 22 3a 20 22 68 74 74 70 3a 2f 2f 68 74 74 70 62 69
#> [326] 6e 2e 6f 72 67 2f 70 6f 73 74 22 0a 7d 0a
#> 
#> $data
#> [1] "{\n  \"args\": {}, \n  \"data\": \"hello world\", \n  \"files\": {}, \n  \"form\": {}, \n  \"headers\": {\n    \"Content-Length\": \"11\", \n    \"Content-Type\": \"text-plain\", \n    \"Host\": \"httpbin.org\", \n    \"X-Amzn-Trace-Id\": \"Root=1-6223c5e4-03915e5a5edbcd9332c9b888\"\n  }, \n  \"json\": null, \n  \"origin\": \"104.215.113.141\", \n  \"url\": \"http://httpbin.org/post\"\n}\n"
#>