ncurl

nano cURL - a minimalist http(s) client.

ncurl(
  url,
  async = FALSE,
  convert = TRUE,
  follow = FALSE,
  method = NULL,
  headers = NULL,
  data = NULL,
  response = NULL,
  timeout = NULL,
  tls = NULL
)

Arguments

url

the URL address.

async

[default FALSE] logical value whether to perform an async request, in which case an 'ncurlAio' is returned instead of a list.

convert

[default TRUE] logical value whether to attempt conversion of the received raw bytes to a character vector. Set to FALSE if downloading non-text data. Supplying a non-logical value will error.

follow

[default FALSE] logical value whether to automatically follow redirects (not applicable for async requests). If FALSE (or async), the redirect address is returned as response header 'Location'. Supplying a non-logical value will error.

method

(optional) the HTTP method (defaults to 'GET' if not specified).

headers

(optional) a named list or character vector specifying the HTTP request headers e.g. list(`Content-Type` = "text/plain") or c(Authorization = "Bearer APIKEY"). Supplying a non-named list or vector will error.

data

(optional) the request data to be submitted.

response

(optional) a character vector or list specifying the response headers to return e.g. c("date", "server") or list("Date", "Server"). These are case-insensitive and will return NULL if not present.

timeout

(optional) integer value in milliseconds after which the transaction times out if not yet complete.

tls

(optional) applicable to secure HTTPS sites only, a client TLS Configuration object created by tls_config. If missing or NULL, certificates are not validated.

Value

Named list of 4 elements:

$status - integer HTTP repsonse status code (200 - OK). Use status_code for a translation of the meaning.

$headers - named list of response headers supplied in 'response', or NULL otherwise. If the status code is within the 300 range, i.e. a redirect, the response header 'Location' is automatically appended to return the redirect address.

$raw - if 'convert' = FALSE, the raw vector of the received resource, or NULL otherwise (use writeBin to save to a file).

$data - if 'convert' = TRUE, the converted character string, or NULL otherwise. This may be further parsed as html, json, xml etc. if required.

Or else, if 'async' = TRUE, an 'ncurlAio' (object of class 'ncurlAio' and 'recvAio') (invisibly).

Examples

ncurl("https://www.r-project.org/",
       convert = FALSE,
       response = c("date", "server"),
       timeout = 1000L)
#> $status
#> [1] 200
#> 
#> $headers
#> $headers$date
#> [1] "Tue, 25 Jul 2023 16:16:23 GMT"
#> 
#> $headers$server
#> [1] "Apache"
#> 
#> 
#> $raw
#>    [1] 3c 21 44 4f 43 54 59 50 45 20 68 74 6d 6c 3e 0a 3c 68 74 6d 6c 20 6c 61
#>   [25] 6e 67 3d 22 65 6e 22 3e 0a 20 20 3c 68 65 61 64 3e 0a 20 20 20 20 3c 6d
#>   [49] 65 74 61 20 63 68 61 72 73 65 74 3d 22 75 74 66 2d 38 22 3e 0a 20 20 20
#>   [73] 20 3c 6d 65 74 61 20 68 74 74 70 2d 65 71 75 69 76 3d 22 58 2d 55 41 2d
#>   [97] 43 6f 6d 70 61 74 69 62 6c 65 22 20 63 6f 6e 74 65 6e 74 3d 22 49 45 3d
#>  [121] 65 64 67 65 22 3e 0a 20 20 20 20 3c 6d 65 74 61 20 6e 61 6d 65 3d 22 76
#>  [145] 69 65 77 70 6f 72 74 22 20 63 6f 6e 74 65 6e 74 3d 22 77 69 64 74 68 3d
#>  [169] 64 65 76 69 63 65 2d 77 69 64 74 68 2c 20 69 6e 69 74 69 61 6c 2d 73 63
#>  [193] 61 6c 65 3d 31 22 3e 0a 20 20 20 20 3c 74 69 74 6c 65 3e 52 3a 20 54 68
#>  [217] 65 20 52 20 50 72 6f 6a 65 63 74 20 66 6f 72 20 53 74 61 74 69 73 74 69
#>  [241] 63 61 6c 20 43 6f 6d 70 75 74 69 6e 67 3c 2f 74 69 74 6c 65 3e 0a 0a 20
#>  [265] 20 20 20 3c 6c 69 6e 6b 20 72 65 6c 3d 22 69 63 6f 6e 22 20 74 79 70 65
#>  [289] 3d 22 69 6d 61 67 65 2f 70 6e 67 22 20 68 72 65 66 3d 22 2f 66 61 76 69
#>  [313] 63 6f 6e 2d 33 32 78 33 32 2e 70 6e 67 22 20 73 69 7a 65 73 3d 22 33 32
#>  [337] 78 33 32 22 20 2f 3e 0a 20 20 20 20 3c 6c 69 6e 6b 20 72 65 6c 3d 22 69
#>  [361] 63 6f 6e 22 20 74 79 70 65 3d 22 69 6d 61 67 65 2f 70 6e 67 22 20 68 72
#>  [385] 65 66 3d 22 2f 66 61 76 69 63 6f 6e 2d 31 36 78 31 36 2e 70 6e 67 22 20
#>  [409] 73 69 7a 65 73 3d 22 31 36 78 31 36 22 20 2f 3e 0a 0a 20 20 20 20 3c 21
#>  [433] 2d 2d 20 42 6f 6f 74 73 74 72 61 70 20 2d 2d 3e 0a 20 20 20 20 3c 6c 69
#>  [457] 6e 6b 20 68 72 65 66 3d 22 2f 63 73 73 2f 62 6f 6f 74 73 74 72 61 70 2e
#>  [481] 6d 69 6e 2e 63 73 73 22 20 72 65 6c 3d 22 73 74 79 6c 65 73 68 65 65 74
#>  [505] 22 3e 0a 20 20 20 20 3c 6c 69 6e 6b 20 68 72 65 66 3d 22 2f 63 73 73 2f
#>  [529] 52 2e 63 73 73 22 20 72 65 6c 3d 22 73 74 79 6c 65 73 68 65 65 74 22 3e
#>  [553] 0a 0a 20 20 20 20 3c 21 2d 2d 20 48 54 4d 4c 35 20 73 68 69 6d 20 61 6e
#>  [577] 64 20 52 65 73 70 6f 6e 64 2e 6a 73 20 66 6f 72 20 49 45 38 20 73 75 70
#>  [601] 70 6f 72 74 20 6f 66 20 48 54 4d 4c 35 20 65 6c 65 6d 65 6e 74 73 20 61
#>  [625] 6e 64 20 6d 65 64 69 61 20 71 75 65 72 69 65 73 20 2d 2d 3e 0a 20 20 20
#>  [649] 20 3c 21 2d 2d 20 57 41 52 4e 49 4e 47 3a 20 52 65 73 70 6f 6e 64 2e 6a
#>  [673] 73 20 64 6f 65 73 6e 27 74 20 77 6f 72 6b 20 69 66 20 79 6f 75 20 76 69
#>  [697] 65 77 20 74 68 65 20 70 61 67 65 20 76 69 61 20 66 69 6c 65 3a 2f 2f 20
#>  [721] 2d 2d 3e 0a 20 20 20 20 3c 21 2d 2d 5b 69 66 20 6c 74 20 49 45 20 39 5d
#>  [745] 3e 0a 20 20 20 20 20 20 3c 73 63 72 69 70 74 20 73 72 63 3d 22 68 74 74
#>  [769] 70 73 3a 2f 2f 6f 73 73 2e 6d 61 78 63 64 6e 2e 63 6f 6d 2f 68 74 6d 6c
#>  [793] 35 73 68 69 76 2f 33 2e 37 2e 32 2f 68 74 6d 6c 35 73 68 69 76 2e 6d 69
#>  [817] 6e 2e 6a 73 22 3e 3c 2f 73 63 72 69 70 74 3e 0a 20 20 20 20 20 20 3c 73
#>  [841] 63 72 69 70 74 20 73 72 63 3d 22 68 74 74 70 73 3a 2f 2f 6f 73 73 2e 6d
#>  [865] 61 78 63 64 6e 2e 63 6f 6d 2f 72 65 73 70 6f 6e 64 2f 31 2e 34 2e 32 2f
#>  [889] 72 65 73 70 6f 6e 64 2e 6d 69 6e 2e 6a 73 22 3e 3c 2f 73 63 72 69 70 74
#>  [913] 3e 0a 20 20 20 20 3c 21 5b 65 6e 64 69 66 5d 2d 2d 3e 0a 20 20 3c 2f 68
#>  [937] 65 61 64 3e 0a 20 20 3c 62 6f 64 79 3e 0a 20 20 20 20 3c 64 69 76 20 63
#>  [961] 6c 61 73 73 3d 22 63 6f 6e 74 61 69 6e 65 72 20 70 61 67 65 22 3e 0a 20
#>  [985] 20 20 20 20 20 3c 64 69 76 20 63 6c 61 73 73 3d 22 72 6f 77 22 3e 0a 20
#> [1009] 20 20 20 20 20 20 20 3c 64 69 76 20 63 6c 61 73 73 3d 22 63 6f 6c 2d 78
#> [1033] 73 2d 31 32 20 63 6f 6c 2d 73 6d 2d 6f 66 66 73 65 74 2d 31 20 63 6f 6c
#> [1057] 2d 73 6d 2d 32 20 73 69 64 65 62 61 72 22 20 72 6f 6c 65 3d 22 6e 61 76
#> [1081] 69 67 61 74 69 6f 6e 22 3e 0a 3c 64 69 76 20 63 6c 61 73 73 3d 22 72 6f
#> [1105] 77 22 3e 0a 3c 64 69 76 20 63 6c 61 73 73 3d 22 63 6f 6c 2d 78 73 2d 36
#> [1129] 20 63 6f 6c 2d 73 6d 2d 31 32 22 3e 0a 3c 70 3e 3c 61 20 68 72 65 66 3d
#> [1153] 22 2f 22 3e 3c 69 6d 67 20 73 72 63 3d 22 2f 52 6c 6f 67 6f 2e 70 6e 67
#> [1177] 22 20 77 69 64 74 68 3d 22 31 30 30 22 20 68 65 69 67 68 74 3d 22 37 38
#> [1201] 22 20 61 6c 74 20 3d 20 22 52 22 20 2f 3e 3c 2f 61 3e 3c 2f 70 3e 0a 3c
#> [1225] 70 3e 3c 73 6d 61 6c 6c 3e 3c 61 20 68 72 65 66 3d 22 2f 22 3e 5b 48 6f
#> [1249] 6d 65 5d 3c 2f 61 3e 3c 2f 73 6d 61 6c 6c 3e 3c 2f 70 3e 0a 3c 68 32 20
#> [1273] 69 64 3d 22 64 6f 77 6e 6c 6f 61 64 22 3e 44 6f 77 6e 6c 6f 61 64 3c 2f
#> [1297] 68 32 3e 0a 3c 70 3e 3c 61 20 68 72 65 66 3d 22 68 74 74 70 73 3a 2f 2f
#> [1321] 63 72 61 6e 2e 72 2d 70 72 6f 6a 65 63 74 2e 6f 72 67 2f 6d 69 72 72 6f
#> [1345] 72 73 2e 68 74 6d 6c 22 3e 43 52 41 4e 3c 2f 61 3e 3c 2f 70 3e 0a 3c 68
#> [1369] 32 20 69 64 3d 22 72 2d 70 72 6f 6a 65 63 74 22 3e 52 20 50 72 6f 6a 65
#> [1393] 63 74 3c 2f 68 32 3e 0a 3c 75 6c 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66
#> [1417] 3d 22 2f 61 62 6f 75 74 2e 68 74 6d 6c 22 3e 41 62 6f 75 74 20 52 3c 2f
#> [1441] 61 3e 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 2f 6c 6f
#> [1465] 67 6f 2f 22 3e 4c 6f 67 6f 3c 2f 61 3e 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c
#> [1489] 61 20 68 72 65 66 3d 22 2f 63 6f 6e 74 72 69 62 75 74 6f 72 73 2e 68 74
#> [1513] 6d 6c 22 3e 43 6f 6e 74 72 69 62 75 74 6f 72 73 3c 2f 61 3e 3c 2f 6c 69
#> [1537] 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 2f 6e 65 77 73 2e 68 74 6d
#> [1561] 6c 22 3e 57 68 61 74 e2 80 99 73 20 4e 65 77 3f 3c 2f 61 3e 3c 2f 6c 69
#> [1585] 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 2f 62 75 67 73 2e 68 74 6d
#> [1609] 6c 22 3e 52 65 70 6f 72 74 69 6e 67 20 42 75 67 73 3c 2f 61 3e 3c 2f 6c
#> [1633] 69 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 2f 63 6f 6e 66 65 72 65
#> [1657] 6e 63 65 73 2f 22 3e 43 6f 6e 66 65 72 65 6e 63 65 73 3c 2f 61 3e 3c 2f
#> [1681] 6c 69 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 2f 73 65 61 72 63 68
#> [1705] 2e 68 74 6d 6c 22 3e 53 65 61 72 63 68 3c 2f 61 3e 3c 2f 6c 69 3e 0a 3c
#> [1729] 6c 69 3e 3c 61 20 68 72 65 66 3d 22 2f 6d 61 69 6c 2e 68 74 6d 6c 22 3e
#> [1753] 47 65 74 20 49 6e 76 6f 6c 76 65 64 3a 20 4d 61 69 6c 69 6e 67 20 4c 69
#> [1777] 73 74 73 3c 2f 61 3e 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66
#> [1801] 3d 22 68 74 74 70 73 3a 2f 2f 63 6f 6e 74 72 69 62 75 74 6f 72 2e 72 2d
#> [1825] 70 72 6f 6a 65 63 74 2e 6f 72 67 2f 22 3e 47 65 74 20 49 6e 76 6f 6c 76
#> [1849] 65 64 3a 20 43 6f 6e 74 72 69 62 75 74 69 6e 67 3c 2f 61 3e 3c 2f 6c 69
#> [1873] 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 68 74 74 70 73 3a 2f 2f 64
#> [1897] 65 76 65 6c 6f 70 65 72 2e 52 2d 70 72 6f 6a 65 63 74 2e 6f 72 67 2f 22
#> [1921] 3e 44 65 76 65 6c 6f 70 65 72 20 50 61 67 65 73 3c 2f 61 3e 3c 2f 6c 69
#> [1945] 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 68 74 74 70 73 3a 2f 2f 62
#> [1969] 6c 6f 67 2e 72 2d 70 72 6f 6a 65 63 74 2e 6f 72 67 2f 22 3e 52 20 42 6c
#> [1993] 6f 67 3c 2f 61 3e 3c 2f 6c 69 3e 0a 3c 2f 75 6c 3e 0a 3c 2f 64 69 76 3e
#> [2017] 0a 3c 64 69 76 20 63 6c 61 73 73 3d 22 63 6f 6c 2d 78 73 2d 36 20 63 6f
#> [2041] 6c 2d 73 6d 2d 31 32 22 3e 0a 3c 68 32 20 69 64 3d 22 72 2d 66 6f 75 6e
#> [2065] 64 61 74 69 6f 6e 22 3e 52 20 46 6f 75 6e 64 61 74 69 6f 6e 3c 2f 68 32
#> [2089] 3e 0a 3c 75 6c 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 2f 66 6f 75
#> [2113] 6e 64 61 74 69 6f 6e 2f 22 3e 46 6f 75 6e 64 61 74 69 6f 6e 3c 2f 61 3e
#> [2137] 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 2f 66 6f 75 6e
#> [2161] 64 61 74 69 6f 6e 2f 62 6f 61 72 64 2e 68 74 6d 6c 22 3e 42 6f 61 72 64
#> [2185] 3c 2f 61 3e 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 2f
#> [2209] 66 6f 75 6e 64 61 74 69 6f 6e 2f 6d 65 6d 62 65 72 73 2e 68 74 6d 6c 22
#> [2233] 3e 4d 65 6d 62 65 72 73 3c 2f 61 3e 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c 61
#> [2257] 20 68 72 65 66 3d 22 2f 66 6f 75 6e 64 61 74 69 6f 6e 2f 64 6f 6e 6f 72
#> [2281] 73 2e 68 74 6d 6c 22 3e 44 6f 6e 6f 72 73 3c 2f 61 3e 3c 2f 6c 69 3e 0a
#> [2305] 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 2f 66 6f 75 6e 64 61 74 69 6f 6e
#> [2329] 2f 64 6f 6e 61 74 69 6f 6e 73 2e 68 74 6d 6c 22 3e 44 6f 6e 61 74 65 3c
#> [2353] 2f 61 3e 3c 2f 6c 69 3e 0a 3c 2f 75 6c 3e 0a 3c 68 32 20 69 64 3d 22 68
#> [2377] 65 6c 70 2d 77 69 74 68 2d 72 22 3e 48 65 6c 70 20 57 69 74 68 20 52 3c
#> [2401] 2f 68 32 3e 0a 3c 75 6c 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 2f
#> [2425] 68 65 6c 70 2e 68 74 6d 6c 22 3e 47 65 74 74 69 6e 67 20 48 65 6c 70 3c
#> [2449] 2f 61 3e 3c 2f 6c 69 3e 0a 3c 2f 75 6c 3e 0a 3c 68 32 20 69 64 3d 22 64
#> [2473] 6f 63 75 6d 65 6e 74 61 74 69 6f 6e 22 3e 44 6f 63 75 6d 65 6e 74 61 74
#> [2497] 69 6f 6e 3c 2f 68 32 3e 0a 3c 75 6c 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65
#> [2521] 66 3d 22 68 74 74 70 73 3a 2f 2f 63 72 61 6e 2e 72 2d 70 72 6f 6a 65 63
#> [2545] 74 2e 6f 72 67 2f 6d 61 6e 75 61 6c 73 2e 68 74 6d 6c 22 3e 4d 61 6e 75
#> [2569] 61 6c 73 3c 2f 61 3e 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66
#> [2593] 3d 22 68 74 74 70 73 3a 2f 2f 63 72 61 6e 2e 72 2d 70 72 6f 6a 65 63 74
#> [2617] 2e 6f 72 67 2f 66 61 71 73 2e 68 74 6d 6c 22 3e 46 41 51 73 3c 2f 61 3e
#> [2641] 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 68 74 74 70 73
#> [2665] 3a 2f 2f 6a 6f 75 72 6e 61 6c 2e 72 2d 70 72 6f 6a 65 63 74 2e 6f 72 67
#> [2689] 22 3e 54 68 65 20 52 20 4a 6f 75 72 6e 61 6c 3c 2f 61 3e 3c 2f 6c 69 3e
#> [2713] 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 2f 64 6f 63 2f 62 69 62 2f 52
#> [2737] 2d 62 6f 6f 6b 73 2e 68 74 6d 6c 22 3e 42 6f 6f 6b 73 3c 2f 61 3e 3c 2f
#> [2761] 6c 69 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 2f 63 65 72 74 69 66
#> [2785] 69 63 61 74 69 6f 6e 2e 68 74 6d 6c 22 3e 43 65 72 74 69 66 69 63 61 74
#> [2809] 69 6f 6e 3c 2f 61 3e 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66
#> [2833] 3d 22 2f 6f 74 68 65 72 2d 64 6f 63 73 2e 68 74 6d 6c 22 3e 4f 74 68 65
#> [2857] 72 3c 2f 61 3e 3c 2f 6c 69 3e 0a 3c 2f 75 6c 3e 0a 3c 68 32 20 69 64 3d
#> [2881] 22 6c 69 6e 6b 73 22 3e 4c 69 6e 6b 73 3c 2f 68 32 3e 0a 3c 75 6c 3e 0a
#> [2905] 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 68 74 74 70 73 3a 2f 2f 77 77 77
#> [2929] 2e 62 69 6f 63 6f 6e 64 75 63 74 6f 72 2e 6f 72 67 22 3e 42 69 6f 63 6f
#> [2953] 6e 64 75 63 74 6f 72 3c 2f 61 3e 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c 61 20
#> [2977] 68 72 65 66 3d 22 68 74 74 70 73 3a 2f 2f 72 2d 66 6f 72 67 65 2e 72 2d
#> [3001] 70 72 6f 6a 65 63 74 2e 6f 72 67 2f 22 3e 52 2d 46 6f 72 67 65 3c 2f 61
#> [3025] 3e 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66 3d 22 68 74 74 70
#> [3049] 73 3a 2f 2f 72 2d 68 75 62 2e 67 69 74 68 75 62 2e 69 6f 2f 72 68 75 62
#> [3073] 2f 22 3e 52 2d 48 75 62 3c 2f 61 3e 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c 61
#> [3097] 20 68 72 65 66 3d 22 2f 67 73 6f 63 2e 68 74 6d 6c 22 3e 47 53 6f 43 3c
#> [3121] 2f 61 3e 3c 2f 6c 69 3e 0a 3c 2f 75 6c 3e 0a 3c 2f 64 69 76 3e 0a 3c 2f
#> [3145] 64 69 76 3e 0a 20 20 20 20 20 20 20 20 3c 2f 64 69 76 3e 0a 20 20 20 20
#> [3169] 20 20 20 20 3c 64 69 76 20 63 6c 61 73 73 3d 22 63 6f 6c 2d 78 73 2d 31
#> [3193] 32 20 63 6f 6c 2d 73 6d 2d 37 22 3e 0a 20 20 20 20 20 20 20 20 3c 68 31
#> [3217] 3e 54 68 65 20 52 20 50 72 6f 6a 65 63 74 20 66 6f 72 20 53 74 61 74 69
#> [3241] 73 74 69 63 61 6c 20 43 6f 6d 70 75 74 69 6e 67 3c 2f 68 31 3e 0a 3c 68
#> [3265] 32 20 69 64 3d 22 67 65 74 74 69 6e 67 2d 73 74 61 72 74 65 64 22 3e 47
#> [3289] 65 74 74 69 6e 67 20 53 74 61 72 74 65 64 3c 2f 68 32 3e 0a 3c 70 3e 52
#> [3313] 20 69 73 20 61 20 66 72 65 65 20 73 6f 66 74 77 61 72 65 20 65 6e 76 69
#> [3337] 72 6f 6e 6d 65 6e 74 20 66 6f 72 20 73 74 61 74 69 73 74 69 63 61 6c 20
#> [3361] 63 6f 6d 70 75 74 69 6e 67 20 61 6e 64 0a 67 72 61 70 68 69 63 73 2e 20
#> [3385] 49 74 20 63 6f 6d 70 69 6c 65 73 20 61 6e 64 20 72 75 6e 73 20 6f 6e 20
#> [3409] 61 20 77 69 64 65 20 76 61 72 69 65 74 79 20 6f 66 20 55 4e 49 58 20 70
#> [3433] 6c 61 74 66 6f 72 6d 73 2c 0a 57 69 6e 64 6f 77 73 20 61 6e 64 20 4d 61
#> [3457] 63 4f 53 2e 20 54 6f 20 3c 73 74 72 6f 6e 67 3e 3c 61 0a 68 72 65 66 3d
#> [3481] 22 68 74 74 70 73 3a 2f 2f 63 72 61 6e 2e 72 2d 70 72 6f 6a 65 63 74 2e
#> [3505] 6f 72 67 2f 6d 69 72 72 6f 72 73 2e 68 74 6d 6c 22 3e 64 6f 77 6e 6c 6f
#> [3529] 61 64 20 52 3c 2f 61 3e 3c 2f 73 74 72 6f 6e 67 3e 2c 0a 70 6c 65 61 73
#> [3553] 65 20 63 68 6f 6f 73 65 20 79 6f 75 72 20 70 72 65 66 65 72 72 65 64 20
#> [3577] 3c 61 0a 68 72 65 66 3d 22 68 74 74 70 73 3a 2f 2f 63 72 61 6e 2e 72 2d
#> [3601] 70 72 6f 6a 65 63 74 2e 6f 72 67 2f 6d 69 72 72 6f 72 73 2e 68 74 6d 6c
#> [3625] 22 3e 43 52 41 4e 20 6d 69 72 72 6f 72 3c 2f 61 3e 2e 3c 2f 70 3e 0a 3c
#> [3649] 70 3e 49 66 20 79 6f 75 20 68 61 76 65 20 71 75 65 73 74 69 6f 6e 73 20
#> [3673] 61 62 6f 75 74 20 52 20 6c 69 6b 65 20 68 6f 77 20 74 6f 20 64 6f 77 6e
#> [3697] 6c 6f 61 64 20 61 6e 64 20 69 6e 73 74 61 6c 6c 20 74 68 65 0a 73 6f 66
#> [3721] 74 77 61 72 65 2c 20 6f 72 20 77 68 61 74 20 74 68 65 20 6c 69 63 65 6e
#> [3745] 73 65 20 74 65 72 6d 73 20 61 72 65 2c 20 70 6c 65 61 73 65 20 72 65 61
#> [3769] 64 20 6f 75 72 20 3c 61 0a 68 72 65 66 3d 22 68 74 74 70 73 3a 2f 2f 63
#> [3793] 72 61 6e 2e 52 2d 70 72 6f 6a 65 63 74 2e 6f 72 67 2f 66 61 71 73 2e 68
#> [3817] 74 6d 6c 22 3e 61 6e 73 77 65 72 73 20 74 6f 20 66 72 65 71 75 65 6e 74
#> [3841] 6c 79 20 61 73 6b 65 64 0a 71 75 65 73 74 69 6f 6e 73 3c 2f 61 3e 20 62
#> [3865] 65 66 6f 72 65 20 79 6f 75 20 73 65 6e 64 20 61 6e 20 65 6d 61 69 6c 2e
#> [3889] 3c 2f 70 3e 0a 3c 68 32 20 69 64 3d 22 6e 65 77 73 22 3e 4e 65 77 73 3c
#> [3913] 2f 68 32 3e 0a 3c 75 6c 3e 0a 3c 6c 69 3e 75 73 65 52 21 20 32 30 32 34
#> [3937] 20 77 69 6c 6c 20 62 65 20 61 20 68 79 62 72 69 64 20 63 6f 6e 66 65 72
#> [3961] 65 6e 63 65 2c 20 74 61 6b 69 6e 67 20 70 6c 61 63 65 20 38 2d 31 31 20
#> [3985] 4a 75 6c 79 20 32 30 32 34 0a 69 6e 20 53 61 6c 7a 62 75 72 67 2c 20 41
#> [4009] 75 73 74 72 69 61 2e 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65 66
#> [4033] 3d 22 68 74 74 70 73 3a 2f 2f 63 72 61 6e 2e 72 2d 70 72 6f 6a 65 63 74
#> [4057] 2e 6f 72 67 2f 73 72 63 2f 62 61 73 65 2f 52 2d 34 22 3e 3c 73 74 72 6f
#> [4081] 6e 67 3e 52 20 76 65 72 73 69 6f 6e 0a 34 2e 33 2e 31 20 28 42 65 61 67
#> [4105] 6c 65 20 53 63 6f 75 74 73 29 3c 2f 73 74 72 6f 6e 67 3e 3c 2f 61 3e 20
#> [4129] 68 61 73 20 62 65 65 6e 20 72 65 6c 65 61 73 65 64 20 6f 6e 20 32 30 32
#> [4153] 33 2d 30 36 2d 31 36 2e 3c 2f 6c 69 3e 0a 3c 6c 69 3e 3c 61 20 68 72 65
#> [4177] 66 3d 22 68 74 74 70 73 3a 2f 2f 63 72 61 6e 2e 72 2d 70 72 6f 6a 65 63
#> [4201] 74 2e 6f 72 67 2f 73 72 63 2f 62 61 73 65 2f 52 2d 34 22 3e 3c 73 74 72
#> [4225] 6f 6e 67 3e 52 20 76 65 72 73 69 6f 6e 0a 34 2e 32 2e 33 20 28 53 68 6f
#> [4249] 72 74 73 74 6f 70 20 42 65 61 67 6c 65 29 3c 2f 73 74 72 6f 6e 67 3e 3c
#> [4273] 2f 61 3e 20 68 61 73 20 62 65 65 6e 20 72 65 6c 65 61 73 65 64 20 6f 6e
#> [4297] 0a 32 30 32 33 2d 30 33 2d 31 35 2e 3c 2f 6c 69 3e 0a 3c 6c 69 3e 59 6f
#> [4321] 75 20 63 61 6e 20 73 75 70 70 6f 72 74 20 74 68 65 20 52 20 46 6f 75 6e
#> [4345] 64 61 74 69 6f 6e 20 77 69 74 68 20 61 20 72 65 6e 65 77 61 62 6c 65 20
#> [4369] 73 75 62 73 63 72 69 70 74 69 6f 6e 20 61 73 20 61 0a 3c 61 20 68 72 65
#> [4393] 66 3d 22 68 74 74 70 73 3a 2f 2f 77 77 77 2e 72 2d 70 72 6f 6a 65 63 74
#> [4417] 2e 6f 72 67 2f 66 6f 75 6e 64 61 74 69 6f 6e 2f 64 6f 6e 61 74 69 6f 6e
#> [4441] 73 2e 68 74 6d 6c 22 3e 73 75 70 70 6f 72 74 69 6e 67 0a 6d 65 6d 62 65
#> [4465] 72 3c 2f 61 3e 3c 2f 6c 69 3e 0a 3c 2f 75 6c 3e 0a 3c 68 32 20 69 64 3d
#> [4489] 22 6e 65 77 73 2d 76 69 61 2d 6d 61 73 74 6f 64 6f 6e 22 3e 4e 65 77 73
#> [4513] 20 76 69 61 20 4d 61 73 74 6f 64 6f 6e 3c 2f 68 32 3e 0a 3c 21 2d 2d 0a
#> [4537] 4d 61 73 74 6f 64 6f 6e 20 77 69 64 67 65 74 20 66 72 6f 6d 20 68 74 74
#> [4561] 70 73 3a 2f 2f 67 69 74 6c 61 62 2e 63 6f 6d 2f 69 64 6f 74 6a 2f 6d 61
#> [4585] 73 74 6f 64 6f 6e 2d 65 6d 62 65 64 2d 66 65 65 64 2d 74 69 6d 65 6c 69
#> [4609] 6e 65 0a 46 69 6c 65 73 20 6d 61 73 74 6f 64 6f 6e 2d 66 65 65 64 2d 74
#> [4633] 69 6d 65 6c 69 6e 65 2e 63 73 73 20 61 6e 64 20 6d 61 73 74 6f 64 6f 6e
#> [4657] 2d 66 65 65 64 2d 74 69 6d 65 6c 69 6e 65 2e 6a 73 20 61 72 65 20 66 72
#> [4681] 6f 6d 20 74 68 69 73 20 73 6f 75 72 63 65 0a 2d 2d 3e 0a 3c 6c 69 6e 6b
#> [4705] 20 72 65 6c 3d 22 73 74 79 6c 65 73 68 65 65 74 22 20 68 72 65 66 3d 22
#> [4729] 6d 61 73 74 6f 64 6f 6e 2d 74 69 6d 65 6c 69 6e 65 2e 63 73 73 22 20 2f
#> [4753] 3e 0a 3c 73 63 72 69 70 74 20 73 72 63 3d 22 6d 61 73 74 6f 64 6f 6e 2d
#> [4777] 74 69 6d 65 6c 69 6e 65 2e 6a 73 22 3e 3c 2f 73 63 72 69 70 74 3e 0a 3c
#> [4801] 6c 69 6e 6b 20 72 65 6c 3d 22 73 74 79 6c 65 73 68 65 65 74 22 20 68 72
#> [4825] 65 66 3d 22 6d 61 73 74 6f 64 6f 6e 2d 74 69 6d 65 6c 69 6e 65 2e 63 73
#> [4849] 73 22 20 2f 3e 0a 3c 73 63 72 69 70 74 20 73 72 63 3d 22 6d 61 73 74 6f
#> [4873] 64 6f 6e 2d 74 69 6d 65 6c 69 6e 65 2e 6a 73 22 3e 3c 2f 73 63 72 69 70
#> [4897] 74 3e 0a 3c 64 69 76 20 63 6c 61 73 73 3d 22 6d 74 2d 74 69 6d 65 6c 69
#> [4921] 6e 65 22 3e 0a 3c 64 69 76 20 69 64 3d 22 6d 74 2d 62 6f 64 79 22 20 63
#> [4945] 6c 61 73 73 3d 22 6d 74 2d 62 6f 64 79 22 20 72 6f 6c 65 3d 22 66 65 65
#> [4969] 64 22 3e 0a 3c 70 72 65 3e 3c 63 6f 64 65 3e 26 6c 74 3b 64 69 76 20 63
#> [4993] 6c 61 73 73 3d 26 71 75 6f 74 3b 6c 6f 61 64 69 6e 67 2d 73 70 69 6e 6e
#> [5017] 65 72 26 71 75 6f 74 3b 26 67 74 3b 26 6c 74 3b 2f 64 69 76 26 67 74 3b
#> [5041] 3c 2f 63 6f 64 65 3e 3c 2f 70 72 65 3e 0a 3c 2f 64 69 76 3e 0a 3c 2f 64
#> [5065] 69 76 3e 0a 3c 68 32 20 69 64 3d 22 73 6f 63 69 61 6c 2d 6d 65 64 69 61
#> [5089] 22 3e 53 6f 63 69 61 6c 20 4d 65 64 69 61 3c 2f 68 32 3e 0a 3c 21 2d 2d
#> [5113] 20 72 65 6c 3d 22 6d 65 22 20 72 65 71 75 69 72 65 64 20 74 6f 20 76 65
#> [5137] 72 69 66 79 20 6f 6e 20 4d 61 73 74 6f 64 6f 6e 20 2d 2d 3e 0a 3c 70 3e
#> [5161] 46 6f 6c 6c 6f 77 20 74 68 65 20 52 20 46 6f 75 6e 64 61 74 69 6f 6e 20
#> [5185] 6f 6e 0a 3c 61 20 72 65 6c 3d 22 6d 65 22 20 68 72 65 66 3d 22 68 74 74
#> [5209] 70 73 3a 2f 2f 66 6f 73 73 74 6f 64 6f 6e 2e 6f 72 67 2f 40 52 5f 46 6f
#> [5233] 75 6e 64 61 74 69 6f 6e 22 3e 4d 61 73 74 6f 64 6f 6e 3c 2f 61 3e 2c 0a
#> [5257] 3c 61 20 68 72 65 66 3d 22 68 74 74 70 73 3a 2f 2f 74 77 69 74 74 65 72
#> [5281] 2e 63 6f 6d 2f 5f 52 5f 46 6f 75 6e 64 61 74 69 6f 6e 22 3e 54 77 69 74
#> [5305] 74 65 72 3c 2f 61 3e 2c 20 6f 72 0a 3c 61 20 68 72 65 66 3d 22 68 74 74
#> [5329] 70 73 3a 2f 2f 77 77 77 2e 6c 69 6e 6b 65 64 69 6e 2e 63 6f 6d 2f 63 6f
#> [5353] 6d 70 61 6e 79 2f 74 68 65 2d 72 2d 66 6f 75 6e 64 61 74 69 6f 6e 2d 66
#> [5377] 6f 72 2d 73 74 61 74 69 73 74 69 63 61 6c 2d 63 6f 6d 70 75 74 69 6e 67
#> [5401] 22 3e 4c 69 6e 6b 65 64 49 6e 3c 2f 61 3e 2e 3c 2f 70 3e 0a 3c 21 2d 2d
#> [5425] 2d 20 28 42 6f 69 6c 65 72 70 6c 61 74 65 20 66 6f 72 20 72 65 6c 65 61
#> [5449] 73 65 20 72 75 6e 2d 69 6e 29 0a 2d 20 20 20 5b 2a 2a 52 20 76 65 72 73
#> [5473] 69 6f 6e 20 33 2e 31 2e 33 20 28 53 6d 6f 6f 74 68 20 53 69 64 65 77 61
#> [5497] 6c 6b 29 20 70 72 65 72 65 6c 65 61 73 65 20 76 65 72 73 69 6f 6e 73 2a
#> [5521] 2a 5d 28 68 74 74 70 73 3a 2f 2f 63 72 61 6e 2e 72 2d 70 72 6f 6a 65 63
#> [5545] 74 2e 6f 72 67 2f 73 72 63 2f 62 61 73 65 2d 70 72 65 72 65 6c 65 61 73
#> [5569] 65 2f 29 20 77 69 6c 6c 20 61 70 70 65 61 72 20 73 74 61 72 74 69 6e 67
#> [5593] 20 46 65 62 72 75 61 72 79 20 32 38 2e 20 46 69 6e 61 6c 20 72 65 6c 65
#> [5617] 61 73 65 20 69 73 20 73 63 68 65 64 75 6c 65 64 20 66 6f 72 20 32 30 31
#> [5641] 35 2d 30 33 2d 30 39 2e 0a 2d 2d 3e 0a 20 20 20 20 20 20 20 20 3c 2f 64
#> [5665] 69 76 3e 0a 20 20 20 20 20 20 3c 2f 64 69 76 3e 0a 20 20 20 20 20 20 3c
#> [5689] 64 69 76 20 63 6c 61 73 73 3d 22 72 61 77 20 66 6f 6f 74 65 72 22 3e 0a
#> [5713] 20 20 20 20 20 20 20 20 26 63 6f 70 79 3b 20 54 68 65 20 52 20 46 6f 75
#> [5737] 6e 64 61 74 69 6f 6e 2e 20 46 6f 72 20 71 75 65 72 69 65 73 20 61 62 6f
#> [5761] 75 74 20 74 68 69 73 20 77 65 62 20 73 69 74 65 2c 20 70 6c 65 61 73 65
#> [5785] 20 63 6f 6e 74 61 63 74 0a 09 3c 73 63 72 69 70 74 20 74 79 70 65 3d 27
#> [5809] 74 65 78 74 2f 6a 61 76 61 73 63 72 69 70 74 27 3e 0a 3c 21 2d 2d 0a 76
#> [5833] 61 72 20 73 3d 22 3d 62 21 69 73 66 67 3e 23 6e 62 6a 6d 75 70 3b 78 66
#> [5857] 63 6e 62 74 75 66 73 41 73 2e 71 73 70 6b 66 64 75 2f 70 73 68 23 3f 75
#> [5881] 69 66 21 78 66 63 6e 62 74 75 66 73 3d 30 62 3f 22 3b 0a 6d 3d 22 22 3b
#> [5905] 20 66 6f 72 20 28 69 3d 30 3b 20 69 3c 73 2e 6c 65 6e 67 74 68 3b 20 69
#> [5929] 2b 2b 29 20 7b 69 66 28 73 2e 63 68 61 72 43 6f 64 65 41 74 28 69 29 20
#> [5953] 3d 3d 20 32 38 29 7b 6d 2b 3d 20 27 26 27 3b 7d 20 65 6c 73 65 20 69 66
#> [5977] 20 28 73 2e 63 68 61 72 43 6f 64 65 41 74 28 69 29 20 3d 3d 20 32 33 29
#> [6001] 20 7b 6d 2b 3d 20 27 21 27 3b 7d 20 65 6c 73 65 20 7b 6d 2b 3d 53 74 72
#> [6025] 69 6e 67 2e 66 72 6f 6d 43 68 61 72 43 6f 64 65 28 73 2e 63 68 61 72 43
#> [6049] 6f 64 65 41 74 28 69 29 2d 31 29 3b 7d 7d 64 6f 63 75 6d 65 6e 74 2e 77
#> [6073] 72 69 74 65 28 6d 29 3b 2f 2f 2d 2d 3e 0a 09 3c 2f 73 63 72 69 70 74 3e
#> [6097] 3b 0a 20 20 20 20 20 20 20 20 66 6f 72 20 71 75 65 72 69 65 73 20 61 62
#> [6121] 6f 75 74 20 52 20 69 74 73 65 6c 66 2c 20 70 6c 65 61 73 65 20 63 6f 6e
#> [6145] 73 75 6c 74 20 74 68 65 20 0a 20 20 20 20 20 20 20 20 3c 61 20 68 72 65
#> [6169] 66 3d 22 68 65 6c 70 2e 68 74 6d 6c 22 3e 47 65 74 74 69 6e 67 20 48 65
#> [6193] 6c 70 3c 2f 61 3e 20 73 65 63 74 69 6f 6e 2e 0a 20 20 20 20 20 20 3c 2f
#> [6217] 64 69 76 3e 0a 20 20 20 20 3c 2f 64 69 76 3e 0a 20 20 20 20 3c 21 2d 2d
#> [6241] 20 6a 51 75 65 72 79 20 28 6e 65 63 65 73 73 61 72 79 20 66 6f 72 20 42
#> [6265] 6f 6f 74 73 74 72 61 70 27 73 20 4a 61 76 61 53 63 72 69 70 74 20 70 6c
#> [6289] 75 67 69 6e 73 29 20 2d 2d 3e 0a 20 20 20 20 3c 73 63 72 69 70 74 20 73
#> [6313] 72 63 3d 22 68 74 74 70 73 3a 2f 2f 61 6a 61 78 2e 67 6f 6f 67 6c 65 61
#> [6337] 70 69 73 2e 63 6f 6d 2f 61 6a 61 78 2f 6c 69 62 73 2f 6a 71 75 65 72 79
#> [6361] 2f 31 2e 31 31 2e 31 2f 6a 71 75 65 72 79 2e 6d 69 6e 2e 6a 73 22 3e 3c
#> [6385] 2f 73 63 72 69 70 74 3e 0a 20 20 20 20 3c 21 2d 2d 20 49 6e 63 6c 75 64
#> [6409] 65 20 61 6c 6c 20 63 6f 6d 70 69 6c 65 64 20 70 6c 75 67 69 6e 73 20 28
#> [6433] 62 65 6c 6f 77 29 2c 20 6f 72 20 69 6e 63 6c 75 64 65 20 69 6e 64 69 76
#> [6457] 69 64 75 61 6c 20 66 69 6c 65 73 20 61 73 20 6e 65 65 64 65 64 20 2d 2d
#> [6481] 3e 0a 20 20 20 20 3c 73 63 72 69 70 74 20 73 72 63 3d 22 2f 6a 73 2f 62
#> [6505] 6f 6f 74 73 74 72 61 70 2e 6d 69 6e 2e 6a 73 22 3e 3c 2f 73 63 72 69 70
#> [6529] 74 3e 0a 20 20 3c 2f 62 6f 64 79 3e 0a 3c 2f 68 74 6d 6c 3e 0a
#> 
#> $data
#> NULL
#> 
ncurl("https://postman-echo.com/put",
      method = "PUT",
      headers = list(Authorization = "Bearer APIKEY"),
      data = "hello world",
      timeout = 1500L)
#> $status
#> [1] 200
#> 
#> $headers
#> NULL
#> 
#> $raw
#> NULL
#> 
#> $data
#> [1] "{\n  \"args\": {},\n  \"data\": \"hello world\",\n  \"files\": {},\n  \"form\": {},\n  \"headers\": {\n    \"x-forwarded-proto\": \"https\",\n    \"x-forwarded-port\": \"443\",\n    \"host\": \"postman-echo.com\",\n    \"x-amzn-trace-id\": \"Root=1-64bff557-76f505b32dd6eb6e42f085ee\",\n    \"content-length\": \"11\",\n    \"authorization\": \"Bearer APIKEY\",\n    \"content-type\": \"application/json\"\n  },\n  \"json\": null,\n  \"url\": \"https://postman-echo.com/put\"\n}"
#> 
ncurl("https://postman-echo.com/post",
      method = "POST",
      headers = c(`Content-Type` = "application/json"),
      data = '{"k":"v"}',
      timeout = 1500L)
#> $status
#> [1] 200
#> 
#> $headers
#> NULL
#> 
#> $raw
#> NULL
#> 
#> $data
#> [1] "{\n  \"args\": {},\n  \"data\": {\n    \"k\": \"v\"\n  },\n  \"files\": {},\n  \"form\": {},\n  \"headers\": {\n    \"x-forwarded-proto\": \"https\",\n    \"x-forwarded-port\": \"443\",\n    \"host\": \"postman-echo.com\",\n    \"x-amzn-trace-id\": \"Root=1-64bff557-22771bec5d2af435094668ec\",\n    \"content-length\": \"9\",\n    \"content-type\": \"application/json\"\n  },\n  \"json\": {\n    \"k\": \"v\"\n  },\n  \"url\": \"https://postman-echo.com/post\"\n}"
#>