Random Data Generation

Strictly not for use in statistical analysis. Non-reproducible and with unknown statistical properties. Provides an alternative source of randomness from the Mbed TLS library for purposes such as cryptographic key generation. Mbed TLS uses a block-cipher in counter mode operation, as defined in NIST SP800-90A: Recommendation for Random Number Generation Using Deterministic Random Bit Generators. The implementation uses AES-256 as the underlying block cipher, with a derivation function, and an entropy collector combining entropy from multiple sources including at least one strong entropy source.

random(n = 1L, convert = TRUE)

Arguments

n

[default 1L] integer random bytes to generate.

convert

[default TRUE] logical FALSE to return a raw vector, or TRUE to return the hex representation of the bytes as a character string.

Value

A length 'n' raw vector, or length one vector of '2n' random characters, depending on the value of 'convert' supplied.

Details

If 'n' is non-integer, it will be coerced to integer; if a vector, only the first element will be used.

Note

Results obtained are independent of and do not alter the state of R's own pseudo-random number generators.

Examples

random()
#> [1] "90"
random(8L)
#> [1] "75ab01f500dd7567"
random(n = 8L, convert = FALSE)
#> [1] db 43 3b 39 11 99 d0 59