Refhook Constructor

Create a function with the required signature for the 'refhook' argument of socket.

refhook(valid, encode, decode)

Arguments

valid

validation code e.g. inherits(x, "torch_tensor") or typeof(x) == "externalptr".

encode

the encode function returning a raw vector e.g. torch:::torch_serialize(x).

decode

the decode function from a raw vector back to the original object e.g. torch::torch_load(x).

Value

A function.

Details

The expressions supplied should all reference the single argument 'x'.

'valid' will be evaluated within an if() block and 'encode' and 'decode' should be wrapped in {} if a compound statement.

Examples

refhook(typeof(x) == "weakref", base64enc(weakref_value(x), convert = FALSE), base64dec(x))
#> function (x) 
#> if (typeof(x) == "weakref") base64enc(weakref_value(x), convert = FALSE) else if (is.raw(x)) base64dec(x)