Set and Retrieve Refhook Function

Sets the refhook function stored internally, used in all message send and receive operations. Calling without an argument retrieves the stored value.

refhook(x = `=`)

Arguments

x

a function or function encoded by base64enc, or NULL to reset.

Value

The current stored refhook function.

Details

The provided function maps to the 'refhook' argument of serialize or unserialize, as the case may be. As the same function is used in both cases, it should contain the logic to distinguish between them as necessary.

Examples

refhook()
#> NULL

func <- function(x) if (is_nano(x)) base64enc(x) else
    if (is.character(x)) unserialize(base64dec(x, convert = FALSE))

refhook(func)
#> function(x) if (is_nano(x)) base64enc(x) else
#>     if (is.character(x)) unserialize(base64dec(x, convert = FALSE))
#> <environment: 0x5566bee4cfe0>

enc <- base64enc(func)
enc
#> [1] "WAoAAAADAAQDAQADBQAAAAAFVVRGLTgAAAYDAAAEAgAAAAEABAAJAAAABnNyY3JlZgAAAw0AAAAIAAAAAwAAAAkAAAAEAAAAQwAAAAkAAABDAAAAAwAAAAQAAAQCAAAAAQAEAAkAAAAHc3JjZmlsZQAAAAQAAAAAAAAA8gAAAP4AAAATAAAAHQAAAP4AAAD+AAAA/gAABAIAAAABAAQACQAAAAJ3ZAAAABAAAAABAAQACQAAACovdG1wL1J0bXBkSDI1TnovZmlsZTFjNDkxNDk5NjRkMi9yZWZlcmVuY2UAAAD+AAAA/gAABAIAAAABAAQACQAAAAVsaW5lcwAAABAAAAAQAAQACQAAAAlyZWZob29rKCkABAAJAAAAAAAEAAkAAAA1ZnVuYyA8LSBmdW5jdGlvbih4KSBpZiAoaXNfbmFubyh4KSkgYmFzZTY0ZW5jKHgpIGVsc2UABAAJAAAAQyAgICBpZiAoaXMuY2hhcmFjdGVyKHgpKSB1bnNlcmlhbGl6ZShiYXNlNjRkZWMoeCwgY29udmVydCA9IEZBTFNFKSkABAAJAAAAAAAEAAkAAAANcmVmaG9vayhmdW5jKQAEAAkAAAAAAAQACQAAABZlbmMgPC0gYmFzZTY0ZW5jKGZ1bmMpAAQACQAAAANlbmMABAAJAAAADHJlZmhvb2soZW5jKQAEAAkAAAAAAAQACQAAAAlyZWZob29rKCkABAAJAAAAAAAEAAkAAAANcmVmaG9vayhOVUxMKQAEAAkAAAAAAAQACQAAAAAAAAD+AAAEAgAAAAEABAAJAAAAA0VuYwAAABAAAAABAAQACQAAAAd1bmtub3duAAAA/gAAAP4AAAQCAAAAAQAEAAkAAAAGaXNGaWxlAAAACgAAAAEAAAAAAAAA/gAABAIAAAABAAQACQAAAAl0aW1lc3RhbXAAAAMOAAAAAUHZJRxkM6MFAAAEAgAAAAEABAAJAAAABWNsYXNzAAAAEAAAAAIABAAJAAAAB1BPU0lYY3QABAAJAAAABlBPU0lYdAAAAP4AAAD+AAAA/gAAAP4AAAD+AAAEAgAAAAEABAAJAAAACGZpbGVuYW1lAAAAEAAAAAEABAAJAAAABjx0ZXh0PgAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAEAgAAAAEABAAJAAAADWZpeGVkTmV3bGluZXMAAAAKAAAAAQAAAAEAAAD+AAAEAgAAAAEABAAJAAAACXBhcnNlRGF0YQAAAw0AAAMgAAAAAQAAAAEAAAABAAAABwAAAAEAAAEoAAAAAQAAAAMAAAABAAAACAAAAAEAAAAIAAAAAQAAACgAAAACAAAABgAAAAEAAAABAAAAAQAAAAcAAAAAAAAATwAAAAMAAAAGAAAAAQAAAAkAAAABAAAACQAAAAEAAAApAAAABAAAAAYAAAABAAAAAQAAAAEAAAAJAAAAAAAAAE8AAAAGAAAAAAAAAAMAAAABAAAAAwAAAAQAAAABAAABBwAAAAwAAAAOAAAAAwAAAAYAAAADAAAABwAAAAEAAAEKAAAADQAAAGAAAAADAAAAAQAAAAMAAAAEAAAAAAAAAE8AAAAOAAAAYAAAAAMAAAAJAAAAAwAAABAAAAABAAABCAAAAA8AAABfAAAAAwAAABEAAAADAAAAEQAAAAEAAAAoAAAAEAAAAF8AAAADAAAAEgAAAAMAAAASAAAAAQAAASQAAAARAAAAXwAAAAMAAAATAAAAAwAAABMAAAABAAAAKQAAABIAAABfAAAAAwAAABUAAAADAAAAFgAAAAEAAAEQAAAAFAAAAF0AAAADAAAAGAAAAAMAAAAYAAAAAQAAACgAAAAVAAAAXQAAAAMAAAAZAAAAAwAAAB8AAAABAAABKAAAABYAAAAYAAAAAwAAACAAAAADAAAAIAAAAAEAAAAoAAAAFwAAAB8AAAADAAAAGQAAAAMAAAAfAAAAAAAAAE8AAAAYAAAAHwAAAAMAAAAhAAAAAwAAACEAAAABAAABBwAAABkAAAAbAAAAAwAAACIAAAADAAAAIgAAAAEAAAApAAAAGgAAAB8AAAADAAAAIQAAAAMAAAAhAAAAAAAAAE8AAAAbAAAAHwAAAAMAAAAZAAAAAwAAACIAAAAAAAAATwAAAB8AAABdAAAAAwAAACMAAAADAAAAIwAAAAEAAAApAAAAIAAAAF0AAAADAAAAJQAAAAMAAAAtAAAAAQAAASgAAAAjAAAAJQAAAAMAAAAuAAAAAwAAAC4AAAABAAAAKAAAACQAAAAsAAAAAwAAACUAAAADAAAALQAAAAAAAABPAAAAJQAAACwAAAADAAAALwAAAAMAAAAvAAAAAQAAAQcAAAAmAAAAKAAAAAMAAAAwAAAAAwAAADAAAAABAAAAKQAAACcAAAAsAAAAAwAAAC8AAAADAAAALwAAAAAAAABPAAAAKAAAACwAAAADAAAAJQAAAAMAAAAwAAAAAAAAAE8AAAAsAAAAXQAAAAMAAAAyAAAAAwAAADUAAAABAAABEQAAAC0AAABdAAAABAAAAAUAAAAEAAAABgAAAAEAAAEQAAAAMAAAAFsAAAAEAAAACAAAAAQAAAAIAAAAAQAAACgAAAAxAAAAWwAAAAQAAAAJAAAABAAAABQAAAABAAABKAAAADIAAAA0AAAABAAAABUAAAAEAAAAFQAAAAEAAAAoAAAAMwAAADsAAAAEAAAACQAAAAQAAAAUAAAAAAAAAE8AAAA0AAAAOwAAAAQAAAAWAAAABAAAABYAAAABAAABBwAAADUAAAA3AAAABAAAABcAAAAEAAAAFwAAAAEAAAApAAAANgAAADsAAAAEAAAAFgAAAAQAAAAWAAAAAAAAAE8AAAA3AAAAOwAAAAQAAAAJAAAABAAAABcAAAAAAAAATwAAADsAAABbAAAABAAAABgAAAAEAAAAGAAAAAEAAAApAAAAPAAAAFsAAAAEAAAAGgAAAAQAAAAkAAAAAQAAASgAAAA/AAAAQQAAAAQAAAAlAAAABAAAACUAAAABAAAAKAAAAEAAAABYAAAABAAAABoAAAAEAAAAJAAAAAAAAABPAAAAQQAAAFgAAAAEAAAAJgAAAAQAAAAuAAAAAQAAASgAAABCAAAARAAAAAQAAAAvAAAABAAAAC8AAAABAAAAKAAAAEMAAABTAAAABAAAACYAAAAEAAAALgAAAAAAAABPAAAARAAAAFMAAAAEAAAAMAAAAAQAAAAwAAAAAQAAAQcAAABFAAAARwAAAAQAAAAxAAAABAAAADEAAAABAAAALAAAAEYAAABTAAAABAAAADAAAAAEAAAAMAAAAAAAAABPAAAARwAAAFMAAAAEAAAAMwAAAAQAAAA5AAAAAQAAAScAAABLAAAAUwAAAAQAAAA7AAAABAAAADsAAAABAAABJgAAAEwAAABTAAAABAAAAD0AAAAEAAAAQQAAAAEAAAEFAAAATQAAAE4AAAAEAAAAPQAAAAQAAABBAAAAAAAAAE8AAABOAAAAUwAAAAQAAABCAAAABAAAAEIAAAABAAAAKQAAAE8AAABTAAAABAAAACYAAAAEAAAAQgAAAAAAAABPAAAAUwAAAFgAAAAEAAAAQwAAAAQAAABDAAAAAQAAACkAAABUAAAAWAAAAAQAAAAaAAAABAAAAEMAAAAAAAAATwAAAFgAAABbAAAABAAAAAUAAAAEAAAAQwAAAAAAAABPAAAAWwAAAF0AAAADAAAAFQAAAAQAAABDAAAAAAAAAE8AAABdAAAAXwAAAAMAAAAJAAAABAAAAEMAAAAAAAAATwAAAF8AAABgAAAAAwAAAAEAAAAEAAAAQwAAAAAAAABPAAAAYAAAAAAAAAAGAAAAAQAAAAYAAAAHAAAAAQAAASgAAABlAAAAZwAAAAYAAAAIAAAABgAAAAgAAAABAAAAKAAAAGYAAABuAAAABgAAAAEAAAAGAAAABwAAAAAAAABPAAAAZwAAAG4AAAAGAAAACQAAAAYAAAAMAAAAAQAAAQcAAABoAAAAagAAAAYAAAANAAAABgAAAA0AAAABAAAAKQAAAGkAAABuAAAABgAAAAkAAAAGAAAADAAAAAAAAABPAAAAagAAAG4AAAAGAAAAAQAAAAYAAAANAAAAAAAAAE8AAABuAAAAAAAAAAgAAAABAAAACAAAAAMAAAABAAABBwAAAHQAAAB2AAAACAAAAAUAAAAIAAAABgAAAAEAAAEKAAAAdQAAAIIAAAAIAAAAAQAAAAgAAAADAAAAAAAAAE8AAAB2AAAAggAAAAgAAAAIAAAACAAAABAAAAABAAABKAAAAHcAAAB5AAAACAAAABEAAAAIAAAAEQAAAAEAAAAoAAAAeAAAAIAAAAAIAAAACAAAAAgAAAAQAAAAAAAAAE8AAAB5AAAAgAAAAAgAAAASAAAACAAAABUAAAABAAABBwAAAHoAAAB8AAAACAAAABYAAAAIAAAAFgAAAAEAAAApAAAAewAAAIAAAAAIAAAAEgAAAAgAAAAVAAAAAAAAAE8AAAB8AAAAgAAAAAgAAAAIAAAACAAAABYAAAAAAAAATwAAAIAAAACCAAAACAAAAAEAAAAIAAAAFgAAAAAAAABPAAAAggAAAAAAAAAJAAAAAQAAAAkAAAADAAAAAQAAAQcAAACFAAAAhwAAAAkAAAABAAAACQAAAAMAAAAAAAAATwAAAIcAAAAAAAAACgAAAAEAAAAKAAAABwAAAAEAAAEoAAAAigAAAIwAAAAKAAAACAAAAAoAAAAIAAAAAQAAACgAAACLAAAAkwAAAAoAAAABAAAACgAAAAcAAAAAAAAATwAAAIwAAACTAAAACgAAAAkAAAAKAAAACwAAAAEAAAEHAAAAjQAAAI8AAAAKAAAADAAAAAoAAAAMAAAAAQAAACkAAACOAAAAkwAAAAoAAAAJAAAACgAAAAsAAAAAAAAATwAAAI8AAACTAAAACgAAAAEAAAAKAAAADAAAAAAAAABPAAAAkwAAAAAAAAAMAAAAAQAAAAwAAAAHAAAAAQAAASgAAACZAAAAmwAAAAwAAAAIAAAADAAAAAgAAAABAAAAKAAAAJoAAACeAAAADAAAAAEAAAAMAAAABwAAAAAAAABPAAAAmwAAAJ4AAAAMAAAACQAAAAwAAAAJAAAAAQAAACkAAACcAAAAngAAAAwAAAABAAAADAAAAAkAAAAAAAAATwAAAJ4AAAAAAAAADgAAAAEAAAAOAAAABwAAAAEAAAEoAAAApAAAAKYAAAAOAAAACAAAAA4AAAAIAAAAAQAAACgAAAClAAAArQAAAA4AAAABAAAADgAAAAcAAAAAAAAATwAAAKYAAACtAAAADgAAAAkAAAAOAAAADAAAAAEAAAEGAAAApwAAAKkAAAAOAAAADQAAAA4AAAANAAAAAQAAACkAAACoAAAArQAAAA4AAAAJAAAADgAAAAwAAAAAAAAATwAAAKkAAACtAAAADgAAAAEAAAAOAAAADQAAAAAAAABPAAAArQAAAAAAAAQCAAAAAQAEAAkAAAADZGltAAAADQAAAAIAAAAIAAAAZAAABAIAAAABAAQACQAAAAZ0b2tlbnMAAAAQAAAAZAAEAAkAAAAUU1lNQk9MX0ZVTkNUSU9OX0NBTEwABAAJAAAAAycoJwAEAAkAAAAEZXhwcgAEAAkAAAADJyknAAQACQAAAARleHByAAQACQAAAAZTWU1CT0wABAAJAAAAC0xFRlRfQVNTSUdOAAQACQAAAARleHByAAQACQAAAAhGVU5DVElPTgAEAAkAAAADJygnAAQACQAAAA5TWU1CT0xfRk9STUFMUwAEAAkAAAADJyknAAQACQAAAAJJRgAEAAkAAAADJygnAAQACQAAABRTWU1CT0xfRlVOQ1RJT05fQ0FMTAAEAAkAAAADJygnAAQACQAAAARleHByAAQACQAAAAZTWU1CT0wABAAJAAAAAycpJwAEAAkAAAAEZXhwcgAEAAkAAAAEZXhwcgAEAAkAAAADJyknAAQACQAAABRTWU1CT0xfRlVOQ1RJT05fQ0FMTAAEAAkAAAADJygnAAQACQAAAARleHByAAQACQAAAAZTWU1CT0wABAAJAAAAAycpJwAEAAkAAAAEZXhwcgAEAAkAAAAEZXhwcgAEAAkAAAAERUxTRQAEAAkAAAACSUYABAAJAAAAAycoJwAEAAkAAAAUU1lNQk9MX0ZVTkNUSU9OX0NBTEwABAAJAAAAAycoJwAEAAkAAAAEZXhwcgAEAAkAAAAGU1lNQk9MAAQACQAAAAMnKScABAAJAAAABGV4cHIABAAJAAAABGV4cHIABAAJAAAAAycpJwAEAAkAAAAUU1lNQk9MX0ZVTkNUSU9OX0NBTEwABAAJAAAAAycoJwAEAAkAAAAEZXhwcgAEAAkAAAAUU1lNQk9MX0ZVTkNUSU9OX0NBTEwABAAJAAAAAycoJwAEAAkAAAAEZXhwcgAEAAkAAAAGU1lNQk9MAAQACQAAAAMnLCcABAAJAAAABGV4cHIABAAJAAAAClNZTUJPTF9TVUIABAAJAAAABkVRX1NVQgAEAAkAAAAJTlVNX0NPTlNUAAQACQAAAARleHByAAQACQAAAAMnKScABAAJAAAABGV4cHIABAAJAAAAAycpJwAEAAkAAAAEZXhwcgAEAAkAAAAEZXhwcgAEAAkAAAAEZXhwcgAEAAkAAAAEZXhwcgAEAAkAAAAEZXhwcgAEAAkAAAAUU1lNQk9MX0ZVTkNUSU9OX0NBTEwABAAJAAAAAycoJwAEAAkAAAAEZXhwcgAEAAkAAAAGU1lNQk9MAAQACQAAAAMnKScABAAJAAAABGV4cHIABAAJAAAABGV4cHIABAAJAAAABlNZTUJPTAAEAAkAAAALTEVGVF9BU1NJR04ABAAJAAAABGV4cHIABAAJAAAAFFNZTUJPTF9GVU5DVElPTl9DQUxMAAQACQAAAAMnKCcABAAJAAAABGV4cHIABAAJAAAABlNZTUJPTAAEAAkAAAADJyknAAQACQAAAARleHByAAQACQAAAARleHByAAQACQAAAARleHByAAQACQAAAAZTWU1CT0wABAAJAAAABGV4cHIABAAJAAAAFFNZTUJPTF9GVU5DVElPTl9DQUxMAAQACQAAAAMnKCcABAAJAAAABGV4cHIABAAJAAAABlNZTUJPTAAEAAkAAAADJyknAAQACQAAAARleHByAAQACQAAAARleHByAAQACQAAABRTWU1CT0xfRlVOQ1RJT05fQ0FMTAAEAAkAAAADJygnAAQACQAAAARleHByAAQACQAAAAMnKScABAAJAAAABGV4cHIABAAJAAAAFFNZTUJPTF9GVU5DVElPTl9DQUxMAAQACQAAAAMnKCcABAAJAAAABGV4cHIABAAJAAAACk5VTExfQ09OU1QABAAJAAAAAycpJwAEAAkAAAAEZXhwcgAEAAkAAAAEZXhwcgAABAIAAAABAAQACQAAAAR0ZXh0AAAAEAAAAGQABAAJAAAAB3JlZmhvb2sABAAJAAAAASgABAAJAAAAAAAEAAkAAAABKQAEAAkAAAAAAAQACQAAAARmdW5jAAQACQAAAAI8LQAEAAkAAAAAAAQACQAAAAhmdW5jdGlvbgAEAAkAAAABKAAEAAkAAAABeAAEAAkAAAABKQAEAAkAAAACaWYABAAJAAAAASgABAAJAAAAB2lzX25hbm8ABAAJAAAAASgABAAJAAAAAAAEAAkAAAABeAAEAAkAAAABKQAEAAkAAAAAAAQACQAAAAAABAAJAAAAASkABAAJAAAACWJhc2U2NGVuYwAEAAkAAAABKAAEAAkAAAAAAAQACQAAAAF4AAQACQAAAAEpAAQACQAAAAAABAAJAAAAAAAEAAkAAAAEZWxzZQAEAAkAAAACaWYABAAJAAAAASgABAAJAAAADGlzLmNoYXJhY3RlcgAEAAkAAAABKAAEAAkAAAAAAAQACQAAAAF4AAQACQAAAAEpAAQACQAAAAAABAAJAAAAAAAEAAkAAAABKQAEAAkAAAALdW5zZXJpYWxpemUABAAJAAAAASgABAAJAAAAAAAEAAkAAAAJYmFzZTY0ZGVjAAQACQAAAAEoAAQACQAAAAAABAAJAAAAAXgABAAJAAAAASwABAAJAAAAAAAEAAkAAAAHY29udmVydAAEAAkAAAABPQAEAAkAAAAFRkFMU0UABAAJAAAAAAAEAAkAAAABKQAEAAkAAAAAAAQACQAAAAEpAAQACQAAAAAABAAJAAAAAAAEAAkAAAAAAAQACQAAAAAABAAJAAAAAAAEAAkAAAAHcmVmaG9vawAEAAkAAAABKAAEAAkAAAAAAAQACQAAAARmdW5jAAQACQAAAAEpAAQACQAAAAAABAAJAAAAAAAEAAkAAAADZW5jAAQACQAAAAI8LQAEAAkAAAAAAAQACQAAAAliYXNlNjRlbmMABAAJAAAAASgABAAJAAAAAAAEAAkAAAAEZnVuYwAEAAkAAAABKQAEAAkAAAAAAAQACQAAAAAABAAJAAAAAAAEAAkAAAADZW5jAAQACQAAAAAABAAJAAAAB3JlZmhvb2sABAAJAAAAASgABAAJAAAAAAAEAAkAAAADZW5jAAQACQAAAAEpAAQACQAAAAAABAAJAAAAAAAEAAkAAAAHcmVmaG9vawAEAAkAAAABKAAEAAkAAAAAAAQACQAAAAEpAAQACQAAAAAABAAJAAAAB3JlZmhvb2sABAAJAAAAASgABAAJAAAAAAAEAAkAAAAETlVMTAAEAAkAAAABKQAEAAkAAAAAAAQACQAAAAAAAAQCAAAJ/wAAABAAAAABAAQACQAAAAlwYXJzZURhdGEAAAD+AAAA/gAABAIAAAn/AAAAEAAAAAIABAAJAAAAC3NyY2ZpbGVjb3B5AAQACQAAAAdzcmNmaWxlAAAA/gAABAIAAAn/AAAAEAAAAAEABAAJAAAABnNyY3JlZgAAAP4AAAD+AAAABAAAAAAAAAAEAAAAAAAAAAQAAAAAAAAABAAAAAAAAAD9AAAA/gAAABMAAAAdAAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAABMAAAAdAAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAABMAAAAdAAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAABMAAAAdAAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAABAIAAAABAAQACQAAAARmdW5jAAAGAwAABAIAAAH/AAADDQAAAAgAAAADAAAACQAAAAQAAABDAAAACQAAAEMAAAADAAAABAAABAIAAAL/AAAD/wAABAIAAAn/AAAAEAAAAAEABAAJAAAABnNyY3JlZgAAAP4AAAD+AAAQ/wAABAIAAAABAAQACQAAAAF4AAAA+wAAAP4AAAAGAAAAAQAEAAkAAAACaWYAAAACAAAABgAAAAEABAAJAAAAB2lzX25hbm8AAAACAAAV/wAAAP4AAAACAAAABgAAAAEABAAJAAAACWJhc2U2NGVuYwAAAAIAABX/AAAA/gAAAAIAAAAGAAAW/wAAAAIAAAAGAAAAAQAEAAkAAAAMaXMuY2hhcmFjdGVyAAAAAgAAFf8AAAD+AAAAAgAAAAYAAAABAAQACQAAAAt1bnNlcmlhbGl6ZQAAAAIAAAAGAAAAAQAEAAkAAAAJYmFzZTY0ZGVjAAAAAgAAFf8AAAQCAAAAAQAEAAkAAAAHY29udmVydAAAAAoAAAABAAAAAAAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAAAP4AAAD+AAAA/gAABAIAABX/AAAA+wAAAP4AAAAGAAAW/wAAAAIAAAAGAAAX/wAAAAIAABX/AAAA/gAAAAIAAAAGAAAY/wAAAAIAABX/AAAA/gAAAAIAAAAGAAAW/wAAAAIAAAAGAAAZ/wAAAAIAABX/AAAA/gAAAAIAAAAGAAAa/wAAAAIAAAAGAAAb/wAAAAIAABX/AAAEAgAAHP8AAAAKAAAAAQAAAAAAAAD+AAAA/gAAAP4AAAD+"
refhook(enc)
#> function(x) if (is_nano(x)) base64enc(x) else
#>     if (is.character(x)) unserialize(base64dec(x, convert = FALSE))
#> <environment: 0x5566bf0fb540>

refhook()
#> function(x) if (is_nano(x)) base64enc(x) else
#>     if (is.character(x)) unserialize(base64dec(x, convert = FALSE))
#> <environment: 0x5566bf0fb540>

refhook(NULL)
#> NULL