Open Socket

Open a Socket implementing 'protocol', and optionally dial (establish an outgoing connection) or listen (accept an incoming connection) at an address.

socket(
  protocol = c("bus", "pair", "push", "pull", "pub", "sub", "req", "rep", "surveyor",
    "respondent"),
  dial = NULL,
  listen = NULL,
  tls = NULL,
  autostart = TRUE,
  raw = FALSE,
  refhook = NULL
)

Arguments

protocol

[default 'bus'] choose protocol - 'bus', 'pair', 'push', 'pull', 'pub', 'sub', 'req', 'rep', 'surveyor', or 'respondent' - see protocols.

dial

(optional) a URL to dial, specifying the transport and address as a character string e.g. 'inproc://anyvalue' or 'tcp://127.0.0.1:5555' (see transports).

listen

(optional) a URL to listen at, specifying the transport and address as a character string e.g. 'inproc://anyvalue' or 'tcp://127.0.0.1:5555' (see transports).

tls

[default NULL] for secure tls+tcp:// or wss:// connections only, provide a TLS configuration object created by tls_config.

autostart

[default TRUE] whether to start the dialer/listener. Set to FALSE if setting configuration options on the dialer/listener as it is not generally possible to change these once started. For dialers only: set to NA to start synchronously - this is less resilient if a connection is not immediately possible, but avoids subtle errors from attempting to use the socket before an asynchronous dial has completed.

raw

[default FALSE] whether to open raw mode sockets. Note: not for general use - do not enable unless you have a specific need, such as for use with device (refer to NNG documentation).

refhook

[default NULL] (for sending/receiving serialised objects only) register a function to handle non-system reference objects (all external pointers, weak references, and environments other than namespace and package environments and .GlobalEnv). This function must have the signature:

function(x) if ( <validate class of reference object> ) { <encode function - must return a raw vector> } else
if (is.raw(x)) { <decode function> }

The utility refhook may be used to construct such a function. All connected sockets should register the same function (if used) to ensure seamless serialisation / unserialisation.

Value

A Socket (object of class 'nanoSocket' and 'nano').

Details

NNG presents a socket view of networking. The sockets are constructed using protocol-specific functions, as a given socket implements precisely one protocol.

Each socket may be used to send and receive messages (if the protocol supports it, and implements the appropriate protocol semantics). For example, sub sockets automatically filter incoming messages to discard those for topics that have not been subscribed.

This function (optionally) binds a single Dialer and/or Listener to a Socket. More complex network topologies may be created by binding further Dialers/Listeners to the Socket as required using dial and listen.

New contexts may also be created using context if the protocol supports it. Any created contexts will inherit the 'refhook' function registered at the socket.

Note that the argument 'refhook' is an modified version of the 'refhook' argument of serialize and unserialize and not identical. This implementation supports custom serialisation formats which output to a raw vector.

Protocols

The following Scalability Protocols (communication patterns) are implemented:

Bus (mesh networks) - protocol: 'bus'

Pair (two-way radio) - protocol: 'pair'

Pipeline (one-way pipe) - protocol: 'push', 'pull'

Publisher/Subscriber (topics & broadcast) - protocol: 'pub', 'sub'

Request/Reply (RPC) - protocol: 'req', 'rep'

Survey (voting & service discovery) - protocol: 'surveyor', 'respondent'

Please see protocols for further documentation.

Examples

hook <- refhook(typeof(x) == "weakref", base64enc(weakref_value(x), convert = FALSE), base64dec(x))

s <- socket(protocol = "req", listen = "inproc://nanosocket", refhook = hook)
s1 <- socket(protocol = "rep", dial = "inproc://nanosocket", refhook = hook)

wr <- weakref(s, "weakref value")
wr
#> <weak reference>
weakref_value(wr)
#> [1] "weakref value"

send(s, wr)
#> [1] 0
recv(s1)
#> [1] "weakref value"

close(s1)
close(s)